### 2025-10-12 #9192 @editorialbot
[PRE REVIEW]: crystract: A Crystallography Package in R for .cif Data Processing
**Submitting author:** @don-tat (<a href="http://orcid.org/0009-0001-2779-2146">Don Ngo</a>)
**Repository:** https://github.com/PrabhuLab/ml-crystals
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
