### 2025-10-12 937b4a6
Update citation for 'cry' R package
Replaced old citation for 'cry' with updated manual entry.
