+++ paper/paper.bib
@@ -41,11 +41,12 @@
   doi = {10.1107/s2052520616003954}
 }
 
-@misc{Roveda2017,
-  author = {Roveda, J.},
-  title = {{cry: Crystallography in R}},
-  year = {2017},
-  howpublished = {R package version 0.1.2}
+@Manual{,
+  title = {cry: Statistics for Structural Crystallography},
+  author = {James Foadi and David Waterman and Rita Giordano and Kutumbarao Nidamarthi},
+  year = {2025},
+  note = {R package version 0.5.2},
+  url = {https://jfoadi.github.io/cry/},
 }
 
 @article{Wickham2014,
@@ -72,7 +73,7 @@
 @book{Emsley1998,
   author = {Emsley, John},
   title = {The Elements},
-  publisher = {Oxford University Press},
+  publisher = {Oxford University Press},b
   year = {1998}
 }
 
