### 2025-10-12 8fda535
Add new manual entry for cry package
