### 2025-10-12 64edde5
Update citation year and key for Foadi reference
