+++ paper/paper.bib
@@ -41,7 +41,7 @@
   doi = {10.1107/s2052520616003954}
 }
 
-@Manual{Foadi25,
+@Manual{Foadi2025,
   title = {cry: Statistics for Structural Crystallography},
   author = {James Foadi and David Waterman and Rita Giordano and Kutumbarao Nidamarthi},
   year = {2025},
