### 2025-10-12 2340f87
Update citation for the cry package in paper.md
