+++ paper/paper.md
@@ -44,7 +44,7 @@ Other specialized Python tools like `cifkit`—a Python based command-line tool�
 
 Other specialized tools, such as the CCDC's Mercury [@Groom2016] or the IUCr's enCIFer, are indispensable for interactive 3D visualization and formal CIF syntax validation, respectively. However, their primary design as graphical user interface (GUI) applications makes them ill-suited for the automated, scriptable, and reproducible workflows required for modern data science without the possibility of handling a high throughput.
 
-Within the R ecosystem, the landscape is sparse. The `cry` package provides basic statistics for crystallography computation and falls short of large-scale research and ML based applications [@Roveda2017]. Its design, centered on a custom S3 object system, is not optimized for the high-throughput, data-frame-centric workflows required for large-scale statistical analysis. `cry` is limited to the analysis of crystallographic parameters and diffraction data from individual files. It is not equipped for the geometric analysis of atomic structures, as it is unable to apply symmetry operations to generate a full unit cell from asymmetric coordinates; therefore, unable to calculate the resulting interatomic distances and angles, or handle the structural disorder commonly found in real materials.
+Within the R ecosystem, the landscape is sparse. The `cry` package provides basic statistics for crystallography computation and falls short of large-scale research and ML based applications [@Foadi2025]. Its design, centered on a custom S3 object system, is not optimized for the high-throughput, data-frame-centric workflows required for large-scale statistical analysis. `cry` is limited to the analysis of crystallographic parameters and diffraction data from individual files. It is not equipped for the geometric analysis of atomic structures, as it is unable to apply symmetry operations to generate a full unit cell from asymmetric coordinates; therefore, unable to calculate the resulting interatomic distances and angles, or handle the structural disorder commonly found in real materials.
 
 To our knowledge, no package or software—whether in Python, as GUIs, or within R—provides a single, integrated solution capable of combining the automated batch processing of large collections of CIF files and the systematic propagation of experimental uncertainties, while providing additional features indispensable for the structural analysis of large datasets. Such a research software landscape forces researchers into creating fragmented and inefficient workflows, piecing together disparate tools for structure generation, geometric analysis, and finally statistical modeling.
 
