+++ paper/paper.bib
@@ -73,7 +73,7 @@
 @book{Emsley1998,
   author = {Emsley, John},
   title = {The Elements},
-  publisher = {Oxford University Press},b
+  publisher = {Oxford University Press},
   year = {1998}
 }
 
