### 2025-10-15 #9192 @rkurchin
@Don-Tat, yes, absolutely! If you're pushing new commits to the code, no particular action is needed beyond that. If you're updating the paper, you can just run `@editorialbot generate pdf` (needs to be the very start of your comment) to rebuild the paper and ensure the updates appear as you want.
