### 2025-11-10 #9192 @zhubonan
Hello @bobleesj , @samblau and @ml-evs, I would like to invite you to help review this JOSS submission of a CIF processing package. I think it should be very related to your field of research and expertise.

The review process of JOSS is unique as it happens on GitHub. The process is open and transparent. Interactions between the reviewers and the author are done via GitHub issues (and encouraged). You can learn about the reviewing guideline here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

Please let me know if you are able to review this (or not). Thanks in advance.
