### 2025-11-10 #9192 @bobleesj
A quick comment: the paper includes `cifkit` and `CBA in the table. Both are published works:

https://joss.theoj.org/papers/10.21105/joss.07205

https://doi.org/10.1016/j.jallcom.2023.173241

@zhubonan Since some of my work is presented in the paper, would it still be appropriate for me to join as a reviewer?
