### 2025-11-18 655fc44
Fix installation path in README for crystract
