### 2025-11-19 c9d0475
Fix installation path for crystract in README
