+++ packages/crystract/R/process_cif_files.R
@@ -1,3 +1,29 @@
+#' Internal, Safe Worker for Parallel CIF Analysis
+#'
+#' @description A robust wrapper for `analyze_single_cif` that includes error
+#'   handling. Designed to be called by `future.apply` without capturing a
+#'   large parent environment.
+#'
+#' @param cif_content A data.table of CIF content.
+#' @param file_name The name of the file.
+#' @param ... Additional arguments passed directly to `analyze_single_cif`.
+#' @return A one-row data.table on success, or NULL on failure.
+#' @noRd
+.analyze_single_cif_safe <- function(cif_content, file_name, ...) {
+  tryCatch({
+    analyze_single_cif(cif_content = cif_content, file_name = file_name, ...)
+  }, error = function(e) {
+    warning(
+      sprintf(
+        "A critical error occurred while processing '%s': %s",
+        file_name,
+        e$message
+      )
+    )
+    return(NULL)
+  })
+}
+
 #' @title Read CIF Files into Memory
 #' @description Reads one or more CIF files from disk and loads each into a
 #'   `data.table`.
@@ -67,7 +93,7 @@ analyze_single_cif <- function(cif_content,
   space_group_name <- extract_space_group_name(cif_content)
   space_group_number <- extract_space_group_number(cif_content)
   unit_cell_metrics <- extract_unit_cell_metrics(cif_content)
-  atomic_coordinates <- extract_atomic_coordinates(cif_content)
+  atomic_coordinates <- extract_atomic_coordinates(cif_content, chemical_formula)
   symmetry_operations <- extract_symmetry_operations(cif_content)
 
   # --- Step 2: Validate Essential Data for Calculations ---
@@ -103,19 +129,7 @@ analyze_single_cif <- function(cif_content,
         sep = ""
       )
     )
-    return(
-      data.table(
-        file_name = file_name,
-        database_code = database_code,
-        chemical_formula = chemical_formula,
-        structure_type = structure_type,
-        space_group_name = space_group_name,
-        space_group_number = space_group_number,
-        unit_cell_metrics = list(unit_cell_metrics),
-        atomic_coordinates = list(atomic_coordinates),
-        symmetry_operations = list(symmetry_operations)
-      )
-    )
+    return(null)
   }
 
   # --- Step 3: Initialize Result Variables ---
@@ -221,78 +235,238 @@ analyze_single_cif <- function(cif_content,
 }
 
 #' @title Analyze a Batch of CIF Files
-#' @description A high-level wrapper function that reads and analyzes one or more
-#'   CIF files in batch. It calls `analyze_single_cif` for each file.
-#' @param file_paths A character vector of paths to the CIF files to be analyzed.
-#' @param perform_extraction Logical. If `TRUE` (default), extracts all metadata
-#'   and base crystallographic data.
-#' @param perform_calcs_and_transforms Logical. If `TRUE` (default), generates
-#'   the full unit cell, expands it to a supercell, and calculates distances.
-#' @param bonding_algorithms A character vector specifying which bonding algorithms
-#'   to apply. The first algorithm listed is used for subsequent calculations.
-#'   Default: `c("minimum_distance")`. Options: `"minimum_distance"`, `"brunner"`,
-#'   `"hoppe"`. Use `"none"` to skip bonding analysis.
-#' @param calculate_bond_angles Logical. If `TRUE` (default), calculates bond
-#'   angles based on the primary bonding algorithm.
-#' @param perform_error_propagation Logical. If `TRUE` (default), propagates
-#'   experimental uncertainties. Any missing uncertainty values in a CIF file
-#'   will be treated as zero during this step.
-#' @return A `data.table` where each row summarizes the analysis of one CIF file.
[…]
