+++ packages/crystract/R/utils_global.R
@@ -1,177 +1,124 @@
 utils::globalVariables(
   c(
+    # Core data.table symbols
     ".",
     ".N",
-    "V1",
+
+    # Column names used across multiple files
+    "Angle",
+    "AngleError",
     "Atom1",
     "Atom2",
-    "Distance",
-    "dmin",
-    "dcut",
+    "BondStrength",
+    "CentralAtom",
     "DeltaX",
     "DeltaY",
     "DeltaZ",
+    "Distance",
+    "DistanceError",
+    "Element1",
+    "Element2",
+    "file_name",
+    "FullWyckoff",
     "Label",
-    "x_a",
-    "y_b",
-    "z_c",
-    "x_error",
-    "y_error",
-    "z_error",
-    "NeighborCount",
-    "WyckoffMultiplicity",
-    "BondStrength",
-    "CentralAtom",
     "Neighbor1",
     "Neighbor2",
-    "Angle",
-    "dx",
-    "dy",
-    "dz",
-    "s_xf",
-    "s_yf",
-    "s_zf",
-    "s_xf1",
-    "s_yf1",
-    "s_zf1",
-    "Original_Atom2",
-    "s_xf2",
-    "s_yf2",
-    "s_zf2",
-    "term_common",
-    "pd_a",
-    "pd_b",
-    "pd_c",
-    "pd_alpha",
-    "pd_beta",
-    "pd_gamma",
-    "pd_xf1",
-    "pd_yf1",
-    "pd_zf1",
-    "pd_xf2",
-    "pd_yf2",
-    "pd_zf2",
-    "variance",
-    "DistanceError",
-    "AngleError",
-    "p_xc_a",
-    "p_xc_b",
-    "p_xc_c",
-    "p_xc_alpha",
-    "p_xc_beta",
-    "p_xc_gamma",
-    "p_xc_xf",
-    "p_xc_yf",
-    "p_xc_zf",
-    "p_yc_a",
-    "p_yc_b",
-    "p_yc_c",
-    "p_yc_alpha",
-    "p_yc_beta",
-    "p_yc_gamma",
-    "p_yc_xf",
-    "p_yc_yf",
-    "p_yc_zf",
-    "p_zc_a",
-    "p_zc_b",
-    "p_zc_c",
-    "p_zc_alpha",
-    "p_zc_beta",
-    "p_zc_gamma",
-    "p_zc_xf",
-    "p_zc_yf",
-    "p_zc_zf",
-    "s_xc_sq",
-    "s_yc_sq",
-    "s_zc_sq",
-    "xc",
-    "yc",
-    "zc",
-    "i.xc",
-    "i.yc",
-    "i.zc",
-    "i.s_xc_sq",
-    "i.s_yc_sq",
-    "i.s_zc_sq",
-    "Parent_C",
-    "Parent_N1",
-    "Parent_N2",
-    "s_xc1_sq",
-    "s_yc1_sq",
-    "s_zc1_sq",
-    "s_xc2_sq",
-    "s_yc2_sq",
-    "s_zc2_sq",
-    "s_xc3_sq",
-    "s_yc3_sq",
-    "s_zc3_sq",
-    "a_vx",
-    "a_vy",
-    "a_vz",
-    "b_vx",
-    "b_vy",
-    "b_vz",
-    "mag_a",
-    "mag_b",
-    "C_val",
-    "p_C_xc1",
-    "p_C_yc1",
-    "p_C_zc1",
-    "p_C_xc2",
-    "p_C_yc2",
-    "p_C_zc2",
-    "p_C_xc3",
-    "p_C_yc3",
-    "p_C_zc3",
-    "s_C_sq",
-    "s_theta_sq",
-    "xc1",
-    "xc2",
[…]
