+++ packages/crystract/man/calculate_weighted_average_network_distance.Rd
@@ -30,6 +30,7 @@ for site multiplicity and occupancy of the central atoms.
 }
 \seealso{
 Other post-processing: 
+\code{\link{aggregate_batch_results}()},
 \code{\link{export_analysis_to_csv}()},
 \code{\link{filter_ghost_distances}()},
 \code{\link{set_radii_data}()}
