### 2025-12-04 418e383
Add new references for cifkit & cba
