+++ paper/paper.bib
@@ -96,3 +96,26 @@
   doi = {10.26434/chemrxiv-2025-f52qs},
   note = {This content is a preprint and has not been peer-reviewed.}
 }
+
+@article{Lee2024,
+  doi = {10.21105/joss.07205},
+  url = {https://doi.org/10.21105/joss.07205},
+  year = {2024},
+  publisher = {The Open Journal},
+  volume = {9},
+  number = {103},
+  pages = {7205},
+  author = {Lee, Sangjoon and Oliynyk, Anton O.},
+  title = {cifkit: A Python package for coordination geometry and atomic site analysis},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Tyvanchuk2024,
+  author = {Tyvanchuk, Yuriy and Babizhetskyy, Volodymyr and Baran, Stanisław and Szytuła, Andrzej and Smetana, Volodymyr and Lee, Sangjoon and Oliynyk, Anton O. and Mudring, Anja-Verena},
+  title = {The crystal and electronic structure of {RE}$_{23}${Co}$_{6.7}${In}$_{20.3}$ ({RE} = {Gd}–{Tm}, {Lu}): A new structure type based on intergrowth of {AlB}$_2$- and {CsCl}-type related slabs},
+  journal = {Journal of Alloys and Compounds},
+  volume = {976},
+  pages = {173241},
+  year = {2024},
+  doi = {10.1016/j.jallcom.2023.173241}
+}
