### 2025-12-08 dee3a51
Revise descriptions of CIF processing tools
Updated references for specialized Python tools and clarified their functionalities.
