+++ paper/paper.md
@@ -40,7 +40,7 @@ Despite the standardization provided by the CIF format, significant practical ba
 
 A number of excellent software tools have been developed to address some of these challenges, yet a comprehensive review reveals a specific and critical gap. The most mature ecosystem for computational materials science currently resides in Python, where `pymatgen` stands as a powerful and widely adopted standard [@Ong2013]. This extensive library offers robust CIF parsing and a host of advanced analysis functions, including the calculation of interatomic distances and the prediction of bonding environments using sophisticated, data-mining approaches like the CrystalNN algorithm [@Pan2020]. While `pymatgen` offers robust CIF parsing and advanced analysis, it has limitations for a complete, statistically rigorous workflow: it does not natively compute bond angles, nor does it programmatically propagate the experimental uncertainties reported in CIFs for its derived geometric quantities. Furthermore, batch processing requires the user to write custom scripts to loop over files. Filtering presents another challenge: although `pymatgen` includes a function to filter for minimum and maximum distances, they have to be specified one element at a time, one file at a time, by the user. This makes it cumbersome to use `pymatgen` for large scale datasets and studies.
 
-Other specialized Python tools like `cifkit`—a Python based command-line tool—and its companion software, `cif-bond-analyzer` (CBA)—designed for the singular task of exporting bond lists—provide fast, lightweight, and native batch processing for geometric analysis. However, their scope is intentionally limited to interatomic distances and bond analysis; they also do not provide native functions for calculating bond angles or propagating experimental uncertainties.
+Other specialized Python tools like `cifkit`—a Python based command-line tool [@Lee2024]—and its companion software, `cif-bond-analyzer` (CBA)—designed for the singular task of exporting bond lists [@Tyvanchuk2023]—designed for the singular task of exporting bond lists—provide fast, lightweight, and native batch processing for geometric analysis. However, their scope is intentionally limited to interatomic distances and bond analysis; they also do not provide native functions for calculating bond angles or propagating experimental uncertainties.
 
 Other specialized tools, such as the CCDC's Mercury [@Groom2016] or the IUCr's enCIFer, are indispensable for interactive 3D visualization and formal CIF syntax validation, respectively. However, their primary design as graphical user interface (GUI) applications makes them ill-suited for the automated, scriptable, and reproducible workflows required for modern data science without the possibility of handling a high throughput.
 
