### 2025-12-08 8172472
Improve CIF parsing robustness and documentation
Enhances the robustness of CIF file parsing in the crystract package, including safer extraction and cleaning of atomic coordinates, improved symmetry operation parsing, and more resilient file reading with encoding sanitization. Updates documentation and examples for clarity, refines .gitignore and .Rbuildignore, and improves README files to better describe usage and data features. Also adds directory support to analyze_cif_files and improves error handling throughout.
