+++ packages/crystract/R/process_coordinates.R
@@ -1,179 +1,89 @@
-# Internal helper function to parse a single symmetry operation string component
-# (e.g., "x-1/2", "y+z", "-z").
-# It returns a list of two elements: 'W_row' (numeric vector of length 3 for x,y,z coefficients)
-# and 'w_scalar' (numeric scalar for translation).
-#' @noRd
-.parse_sym_op_component <- function(op_str) {
-  W_row <- c(0, 0, 0) # Coefficients for x, y, z
-  w_scalar <- 0
-
-  # Clean spaces and ensure consistent sign representation for fractions
-  op_str <- gsub("\\s+", "", op_str)
-  # Normalize signs for easier splitting, e.g., "x-1/2" -> "x+-1/2"
-  op_str <- gsub("(?<=[^\\d/])-", "+-", op_str, perl = TRUE)
-  # Split into terms like "x", "-y", "+1/2". Handles cases like "1/2+x" too.
-  terms <- stringr::str_split(op_str, "(?<=[^/\\d])(?=[+-])", simplify = TRUE)
-  terms <- terms[terms != ""]
-
-  for (term in terms) {
-    if (term == "") next
-
-    # Handle fractional translation (e.g., +1/2, -1/3) or integer/decimal numbers
-    if (grepl("^[+-]?(\\d+/\\d+|\\d*\\.\\d+|\\d+)$", term)) {
-      w_scalar <- w_scalar + eval(parse(text = term)) # Safely evaluate numerical fractions/constants
-      next
-    }
-
-    # Handle x, y, z terms
-    if (grepl("x", term, fixed = TRUE)) {
-      coeff_str <- gsub("x", "", term, fixed = TRUE)
-      if (coeff_str == "" || coeff_str == "+") W_row[1] <- W_row[1] + 1
-      else if (coeff_str == "-") W_row[1] <- W_row[1] - 1
-      else W_row[1] <- W_row[1] + eval(parse(text = coeff_str)) # Handles "1/2x", "-1/3x"
-    }
-    if (grepl("y", term, fixed = TRUE)) {
-      coeff_str <- gsub("y", "", term, fixed = TRUE)
-      if (coeff_str == "" || coeff_str == "+") W_row[2] <- W_row[2] + 1
-      else if (coeff_str == "-") W_row[2] <- W_row[2] - 1
-      else W_row[2] <- W_row[2] + eval(parse(text = coeff_str))
-    }
-    if (grepl("z", term, fixed = TRUE)) {
-      coeff_str <- gsub("z", "", term, fixed = TRUE)
-      if (coeff_str == "" || coeff_str == "+") W_row[3] <- W_row[3] + 1
-      else if (coeff_str == "-") W_row[3] <- W_row[3] - 1
-      else W_row[3] <- W_row[3] + eval(parse(text = coeff_str))
-    }
-  }
-  list(W_row = W_row, w_scalar = w_scalar)
-}
-
-# Main internal function to parse all symmetry operations into matrices
-#' @noRd
-.parse_symmetry_operations_to_matrices <- function(symmetry_operations_dt) {
-  if (nrow(symmetry_operations_dt) == 0) return(NULL)
-
-  num_sym_ops <- nrow(symmetry_operations_dt)
-  W_matrices <- vector("list", num_sym_ops)
-  w_vectors <- matrix(0, nrow = num_sym_ops, ncol = 3)
-
-  for (i in 1:num_sym_ops) {
-    op_row <- symmetry_operations_dt[i]
-    W_current <- matrix(0, 3, 3)
-    w_current <- c(0, 0, 0)
-
-    # Parse x component
-    parsed_x <- .parse_sym_op_component(op_row$x)
[…]
