+++ packages/crystract/R/extract_cif_data.R
@@ -7,13 +7,9 @@
 #' @return A character string of the cleaned value, or NA if not found.
 #' @noRd
 extract_value <- function(cif_content, pattern, remove_pattern = TRUE) {
-  # Use base R's grepl for more robust pattern matching
   matching_lines <- cif_content[grepl(pattern, V1, fixed = TRUE)]
-
   if (nrow(matching_lines) > 0) {
-    # Take the first matching line
     value <- matching_lines$V1[1]
-
     if (remove_pattern) {
       value <- gsub(pattern, "", value, fixed = TRUE)
     }
@@ -26,23 +22,18 @@ extract_value <- function(cif_content, pattern, remove_pattern = TRUE) {
 }
 
 #' @title Extract Database Code from CIF Content
-#' @description Extracts the database code (e.g., ICSD number) from CIF data.
-#' @param cif_content A `data.table` where each row is a line from a CIF file.
+#' @description Extracts the database code identifier (e.g., from `_database_code_`) from the CIF.
+#' @param cif_content A `data.table` containing the lines of a CIF file.
 #' @return A character string of the database code, or `NA` if not found.
 #' @family extractors
 #' @export
-#' @examples
-#' cif_file <- system.file("extdata", "ICSD422.cif", package = "crystract")
-#' if (file.exists(cif_file)) {
-#'   cif_content <- data.table::fread(cif_file, sep = "\n", header = FALSE)
-#'   extract_database_code(cif_content)
-#' }
 extract_database_code <- function (cif_content) {
   extract_value(cif_content, "_database_code_")
 }
 
 #' @title Extract Chemical Formula from CIF Content
-#' @param cif_content A `data.table` where each row is a line from a CIF file.
+#' @description Extracts the chemical sum formula (e.g., from `_chemical_formula_sum`).
+#' @param cif_content A `data.table` containing the lines of a CIF file.
 #' @return A character string of the chemical formula, or `NA` if not found.
 #' @family extractors
 #' @export
@@ -51,7 +42,8 @@ extract_chemical_formula <- function(cif_content) {
 }
 
 #' @title Extract Structure Type from CIF Content
-#' @param cif_content A `data.table` where each row is a line from a CIF file.
+#' @description Extracts the structure type name (e.g., from `_chemical_name_structure_type`).
+#' @param cif_content A `data.table` containing the lines of a CIF file.
 #' @return A character string of the structure type, or `NA` if not found.
 #' @family extractors
 #' @export
@@ -60,7 +52,8 @@ extract_structure_type <- function(cif_content) {
 }
 
 #' @title Extract Space Group Name from CIF Content
-#' @param cif_content A `data.table` where each row is a line from a CIF file.
+#' @description Extracts the Hermann-Mauguin space group name (e.g., from `_space_group_name_H-M_alt`).
+#' @param cif_content A `data.table` containing the lines of a CIF file.
 #' @return A character string of the space group name, or `NA` if not found.
 #' @family extractors
 #' @export
@@ -69,7 +62,8 @@ extract_space_group_name <- function(cif_content) {
 }
 
 #' @title Extract Space Group Number from CIF Content
-#' @param cif_content A `data.table` where each row is a line from a CIF file.
[…]
