### 2025-12-09 cbee23c
Single-line processing for singular cif file
