### 2025-12-11 5cb70ab
Update install instructions to build vignettes
Added 'build_vignettes = TRUE' to the devtools::install_github command in the README files to ensure vignettes are built during installation from GitHub.
