+++ packages/crystract/README.Rmd
@@ -79,7 +79,7 @@ Once the prerequisites are met, open R or RStudio and run the following commands
 install.packages("devtools")
 
 # Install crystract from the  GitHub repositoy
-devtools::install_github("PrabhuLab/ml-crystals", subdir = "packages/crystract")
+devtools::install_github("PrabhuLab/ml-crystals", subdir = "packages/crystract", build_vignettes = TRUE)
 ```
 
 ### Verifying the Installation
