### 2025-12-11 dfb1b1f
Update vignette with advanced batch processing section
Added a new section detailing high-throughput batch processing using the crystract package, including parallelization with the future package and batch-to-disk mode. Provided example code for running parallel jobs and aggregating results, and clarified usage of analyze_single_cif versus analyze_cif_files.
