### 2025-12-11 181f45a
Fix typo in acknowledgements section
Corrected a minor typo in the acknowledgements section.
