### 2025-12-28 56882a2
Update references and citations in paper
Replaced the Pan2020 reference with Pan2021 in both the bibliography and citations, updated Tyvanchuk2023 to Tyvanchuk2024, and added new references for Mercury (Macrae2006) and enCIFer (Allen2004). Adjusted in-text citations and descriptions to match the updated references.
