+++ paper/paper.bib
@@ -19,15 +19,15 @@
   doi = {10.1016/j.commatsci.2012.10.028}
 }
 
-@article{Pan2020,
-  author = {Pan, Fang and Mi, Wenhao and Gari, R. Sri Krishna and Horton, Matthew},
-  title = {{Crystal structure generative modeling with deep learning}},
-  journal = {Journal of Materials Chemistry A},
-  volume = {8},
-  number = {11},
-  pages = {5576--5585},
-  year = {2020},
-  doi = {10.1039/C9TA01378D}
+@article{Pan2021,
+  author = {Pan, Hillary and Ganose, Alex M. and Horton, Matthew and Aykol, Muratahan and Persson, Kristin A. and Zimmermann, Nils E. R. and Jain, Anubhav},
+  title = {Benchmarking Coordination Number Prediction Algorithms on Inorganic Crystal Structures},
+  journal = {Inorganic Chemistry},
+  volume = {60},
+  number = {3},
+  pages = {1590-1603},
+  year = {2021},
+  doi = {10.1021/acs.inorgchem.0c02996}
 }
 
 @article{Groom2016,
@@ -39,6 +39,28 @@
   pages = {171--179},
   year = {2016},
   doi = {10.1107/s2052520616003954}
+}
+
+@article{Macrae2006,
+  author = {Macrae, Clare F. and Edgington, Paul R. and McCabe, Patrick and Pidcock, Elna and Shields, Greg P. and Taylor, Robin and Towler, Matthew and van de Streek, Jacco},
+  title = {{Mercury: visualization and analysis of crystal structures}},
+  journal = {Journal of Applied Crystallography},
+  year = {2006},
+  volume = {39},
+  number = {3},
+  pages = {453--457},
+  doi = {10.1107/S002188980600731X}
+}
+
+@article{Allen2004,
+  author = {Allen, Frank H. and Johnson, Owen and Shields, Gregory P. and Smith, Barry R. and Towler, Matthew},
+  title = {{CIF applications. XV. enCIFer: a program for viewing, editing and visualizing CIFs}},
+  journal = {Journal of Applied Crystallography},
+  year = {2004},
+  volume = {37},
+  number = {2},
+  pages = {335--338},
+  doi = {10.1107/S0021889804003528}
 }
 
 @Manual{Foadi2025,
