### 2025-12-30 346acfe
Improve CIF parsing to handle indented lines like in PCD/AMCSD
Updated regex patterns and comments throughout the CIF extraction functions to correctly handle indented lines in CIF files. This improves robustness when parsing headers, loops, and data blocks that may be preceded by whitespace.
