### 2025-12-30 dec0769
Double checked bibliography entries
