+++ paper/paper.bib
@@ -1,65 +1,69 @@
 @article{Hall1991,
-  author = {Hall, S. R. and Allen, F. H. and Brown, I. D.},
-  title = {The crystallographic information file ({CIF}): a new standard archive file for crystallography},
-  journal = {Acta Crystallographica Section A: Foundations of Crystallography},
-  volume = {47},
-  number = {6},
-  pages = {655--685},
-  year = {1991},
-  doi = {10.1107/s010876739100531x}
+  title={The crystallographic information file (CIF): a new standard archive file for crystallography},
+  author={Hall, Sydney R and Allen, Frank H and Brown, I David},
+  journal={Foundations of Crystallography},
+  volume={47},
+  number={6},
+  pages={655--685},
+  year={1991},
+  publisher={International Union of Crystallography}
 }
 
 @article{Ong2013,
-  author = {Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Ceder, Gerbrand},
-  title = {{Python Materials Genomics (pymatgen): A robust, open-source Python library for materials analysis}},
-  journal = {Computational Materials Science},
-  volume = {68},
-  pages = {314--319},
-  year = {2013},
-  doi = {10.1016/j.commatsci.2012.10.028}
+  title={Python Materials Genomics (pymatgen): A robust, open-source python library for materials analysis},
+  author={Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L and Persson, Kristin A and Ceder, Gerbrand},
+  journal={Computational Materials Science},
+  volume={68},
+  pages={314--319},
+  year={2013},
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.commatsci.2012.10.028}
 }
 
 @article{Pan2021,
-  author = {Pan, Hillary and Ganose, Alex M. and Horton, Matthew and Aykol, Muratahan and Persson, Kristin A. and Zimmermann, Nils E. R. and Jain, Anubhav},
-  title = {Benchmarking Coordination Number Prediction Algorithms on Inorganic Crystal Structures},
-  journal = {Inorganic Chemistry},
-  volume = {60},
-  number = {3},
-  pages = {1590-1603},
-  year = {2021},
+  title={Benchmarking coordination number prediction algorithms on inorganic crystal structures},
+  author={Pan, Hillary and Ganose, Alex M and Horton, Matthew and Aykol, Muratahan and Persson, Kristin A and Zimmermann, Nils ER and Jain, Anubhav},
+  journal={Inorganic chemistry},
+  volume={60},
+  number={3},
+  pages={1590--1603},
+  year={2021},
+  publisher={ACS Publications},
   doi = {10.1021/acs.inorgchem.0c02996}
 }
 
 @article{Groom2016,
-  author = {Groom, Colin R. and Bruno, Ian J. and Lightfoot, Matthew P. and Ward, Suzanna C.},
-  title = {{The Cambridge Structural Database}},
-  journal = {Acta Crystallographica Section B: Structural Science, Crystal Engineering and Materials},
-  volume = {72},
-  number = {2},
-  pages = {171--179},
-  year = {2016},
+  title={The Cambridge structural database},
+  author={Groom, Colin R and Bruno, Ian J and Lightfoot, Matthew P and Ward, Suzanna C},
+  journal={Structural Science},
+  volume={72},
+  number={2},
+  pages={171--179},
+  year={2016},
[…]
