### 2026-01-06 76b18f0
Add tolerance parameter for distance calculations
Introduced a 'tolerance' parameter to filter out near-zero interatomic distances considered as floating-point noise in calculate_distances and related functions. Updated documentation and function signatures to propagate this parameter, and improved error handling for atomic coordinate extraction. Also added minimum_distance_delta parameter for bonding algorithm flexibility.
