### 2026-01-07 ec8d127
Enhance label normalization and symmetry handling
Improves atomic label normalization and uniqueness in extract_cif_data.R, adds SymOp and translation tracking to symmetry expansion in process_coordinates.R, and updates global variables accordingly. Documentation for analyze_single_cif and analyze_cif_files is updated to include minimum_distance_delta. Minor bugfixes and column compatibility improvements are also included.
