+++ packages/crystract/R/calculate_properties.R
@@ -36,27 +36,27 @@ calculate_distances <- function(atomic_coordinates,
   alpha <- as.numeric(unit_cell_metrics$`_cell_angle_alpha`) * pi / 180
   beta <- as.numeric(unit_cell_metrics$`_cell_angle_beta`) * pi / 180
   gamma <- as.numeric(unit_cell_metrics$`_cell_angle_gamma`) * pi / 180
-  
+
   coords_atomic <- as.matrix(atomic_coordinates[, .(x_a, y_b, z_c)])
   coords_expanded <- as.matrix(expanded_coords[, .(x_a, y_b, z_c)])
   labels_atomic <- atomic_coordinates$Label
   labels_expanded <- expanded_coords$Label
-  
+
   delta_x <- outer(coords_atomic[, 1], coords_expanded[, 1], "-")
   delta_y <- outer(coords_atomic[, 2], coords_expanded[, 2], "-")
   delta_z <- outer(coords_atomic[, 3], coords_expanded[, 3], "-")
-  
+
   cos_alpha <- cos(alpha)
   cos_beta <- cos(beta)
   cos_gamma <- cos(gamma)
-  
+
   r2 <- (a^2 * delta_x^2) + (b^2 * delta_y^2) + (c^2 * delta_z^2) +
     (2 * b * c * cos_alpha * delta_y * delta_z) +
     (2 * c * a * cos_beta * delta_z * delta_x) +
     (2 * a * b * cos_gamma * delta_x * delta_y)
-  
+
   r <- sqrt(r2)
-  
+
   # FIX: stringsAsFactors = FALSE prevents "Ops.factor" warnings downstream
   atom_pairs <- expand.grid(
     Atom1 = labels_atomic,
@@ -64,7 +64,7 @@ calculate_distances <- function(atomic_coordinates,
     KEEP.OUT.ATTRS = TRUE,
     stringsAsFactors = FALSE
   )
-  
+
   distances <- data.table(
     Atom1 = atom_pairs$Atom1,
     Atom2 = atom_pairs$Atom2,
@@ -73,10 +73,10 @@ calculate_distances <- function(atomic_coordinates,
     DeltaY = as.vector(delta_y),
     DeltaZ = as.vector(delta_z)
   )
-  
+
   # Filter out self-interactions or near-zero distances based on tolerance
   distances <- distances[Distance > tolerance]
-  
+
   return(distances)
 }
 
@@ -460,8 +460,10 @@ propagate_angle_error <- function(bond_angles,
                                                                     s_c)^2 + (p_zc_alpha * s_alpha_rad)^2 + (p_zc_beta * s_beta_rad)^2 + (p_zc_gamma *
                                                                                                                                             s_gamma_rad)^2 + (p_zc_xf * x_error)^2 + (p_zc_yf * y_error)^2 + (p_zc_zf *
                                                                                                                                                                                                                 z_error)^2]
-  all_frac_coords <- unique(rbind(atomic_coordinates[, .(Label, x_a, y_b, z_c)], expanded_coords), by =
-                              "Label")
+
+  # Select only compatible columns from expanded_coords to avoid column mismatch errors
+  all_frac_coords <- unique(rbind(atomic_coordinates[, .(Label, x_a, y_b, z_c)], expanded_coords[, .(Label, x_a, y_b, z_c)]), by = "Label")
+
   all_cart_coords <- all_frac_coords[, `:=`(
     xc = a * x_a + b * y_b * cos_g + c * z_c * cos_b,
     yc = b * y_b * sin_g + c * z_c * (cos_a - cos_b * cos_g) / sin_g,
