### 2026-01-15 #9529 @singularitti
The link: https://github.com/PrabhuLab/ml-crystals/packages/crystract is invalid.
