### 2026-01-21 5274a82
Add comprehensive data dictionary to README
Expanded the crystract README with a detailed data dictionary covering the structure of the master analysis object, nested data tables, and reference data. This addition clarifies the output formats and data organization for users and developers.
