+++ packages/crystract/README.Rmd
@@ -43,6 +43,133 @@ knitr::opts_chunk$set(
 
 *   **Batch Processing & Export**: The main `analyze_cif_files()` function is designed to process hundreds of files in a single run, and results can be easily exported to a structured directory of CSV files with `export_analysis_to_csv()`.
 
+### Here is a comprehensive Data Dictionary for the **crystract** package.
+
+It is organized into three sections:
+1.  **Master Analysis Object**: The main output structure returned by `analyze_cif_files()` and `analyze_single_cif()`.
+2.  **Nested Data Tables**: The specific structure of the sub-tables contained within the list-columns of the Master Object.
+3.  **Reference Data**: The structure of auxiliary tables like atomic radii.
+
+---
+
+### 1. Master Analysis Object
+**Function Source:** `analyze_cif_files()`, `analyze_single_cif()`
+**Description:** A `data.table` where each row represents a single processed CIF file. Columns 1–6 contain metadata, while the remaining columns are **list-columns** containing nested `data.table`s with detailed structural data.
+
+| Column Name | Data Type | Description |
+| :--- | :--- | :--- |
+| `file_name` | Character | The name of the processed CIF file (e.g., "ICSD422.cif"). |
+| `database_code` | Character | The unique identifier from the source database (e.g., from `_database_code_`). |
+| `chemical_formula` | Character | The chemical sum formula extracted from the CIF. |
+| `structure_type` | Character | The name of the structure type (e.g., "Perovskite"). |
+| `space_group_name` | Character | Hermann-Mauguin space group symbol (e.g., "P n m a"). |
+| `space_group_number` | Character | International Tables space group number. |
+| `unit_cell_metrics` | List (DT) | Nested table containing lattice parameters (see Section 2.1). |
+| `atomic_coordinates` | List (DT) | Nested table of primary asymmetric atoms (see Section 2.2). |
+| `symmetry_operations` | List (DT) | Nested table of symmetry operators (see Section 2.3). |
+| `transformed_coords` | List (DT) | Nested table of the full unit cell atoms. |
+| `expanded_coords` | List (DT) | Nested table of the supercell (3x3x3) atoms. |
+| `distances` | List (DT) | Nested table of all calculated interatomic distances (see Section 2.4). |
+| `bonded_pairs_minimum_distance` | List (DT) | Nested table of bonds detected via Minimum Distance method (see Section 2.4). |
+| `bonded_pairs_brunner` | List (DT) | Nested table of bonds detected via Brunner's method. |
+| `bonded_pairs_hoppe` | List (DT) | Nested table of bonds detected via Hoppe's method. |
+| `neighbor_counts` | List (DT) | Nested table of coordination numbers (see Section 2.5). |
+| `bond_angles` | List (DT) | Nested table of calculated bond angles (see Section 2.6). |
+
+---
+
+### 2. Nested Data Tables
+
+#### 2.1. Unit Cell Metrics
+**Column in Master Object:** `unit_cell_metrics`
+**Description:** Lattice parameters and their standard uncertainties.
+
+| Column Name | Data Type | Description |
+| :--- | :--- | :--- |
+| `_cell_length_a` | Numeric | Unit cell length $a$ in Angstroms (Å). |
+| `_cell_length_b` | Numeric | Unit cell length $b$ in Angstroms (Å). |
+| `_cell_length_c` | Numeric | Unit cell length $c$ in Angstroms (Å). |
+| `_cell_angle_alpha` | Numeric | Unit cell angle $\alpha$ in degrees. |
[…]
