### 2026-01-21 78ecd68
Add pkgdown site config and GitHub Actions workflow
Introduces pkgdown configuration for documentation site generation, adds a GitHub Actions workflow for automated deployment to GitHub Pages, and updates .Rbuildignore and .gitignore accordingly. The DESCRIPTION and documentation are updated with the site URL and reference links.
