### 2026-01-21 dbf0d65
Update navbar to add Articles section
Replaces the 'News' link in the navbar with an 'Articles' dropdown containing links to the package guide and an example use case. The previous 'News' link is now commented out.
