+++ paper/paper.bib
@@ -86,9 +86,8 @@
 }
 
 @article{Ku1966,
-  author = {Ku, Harry H.},
   title={Notes on the use of propagation of error formulas},
-  author={K$\upsilon$, HH},
+  author = {Ku, Harry H.},
   journal={Journal of Research of the National Bureau of Standards: Engineering and instrumentation. Section C.},
   volume={70},
   number={4},
