### 2026-01-26 d4d6059
Improve symmetry and coordinate precision handling
Added a precision parameter to apply_symmetry_operations and expand_transformed_coords to control rounding of coordinates, improving deduplication and floating point handling. Updated analyze_single_cif to propagate precision based on tolerance, simplified essential data validation, and try forcing minimum_distance to filter non-physical distances before calculations. Updated documentation accordingly.
