+++ packages/crystract/R/process_coordinates.R
@@ -2,10 +2,14 @@
 #' @description Generates all symmetry-equivalent atomic positions within the unit cell.
 #' @param atomic_coordinates A `data.table` of asymmetric atoms.
 #' @param symmetry_operations A `data.table` of operations.
+#' @param precision An integer specifying the number of decimal places used to
+#'   round coordinates when identifying unique atoms. Defaults to 4.
 #' @return A `data.table` containing all unique atomic positions in the unit cell.
 #' @family coordinate processors
 #' @export
-apply_symmetry_operations <- function(atomic_coordinates, symmetry_operations) {
+apply_symmetry_operations <- function(atomic_coordinates,
+                                      symmetry_operations,
+                                      precision = 4) {
   if (is.null(atomic_coordinates) || nrow(atomic_coordinates) == 0) {
     return(
       data.table(
@@ -18,38 +22,29 @@ apply_symmetry_operations <- function(atomic_coordinates, symmetry_operations) {
     )
   }
 
-  # Helper to safely evaluate the string math (e.g. "x+1/2")
-  # Uses strict formatting to prevent parsing errors
   safe_eval <- function(op_str, x, y, z) {
-    # Replace coordinates with their values wrapped in parens to preserve order of operations
-    # e.g., "-x" where x is -0.1 becomes "-(-0.1)" instead of "--0.1"
-    expr_str <- gsub("x", sprintf("(%.10f)", x), op_str, fixed = TRUE)
-    expr_str <- gsub("y", sprintf("(%.10f)", y), expr_str, fixed = TRUE)
-    expr_str <- gsub("z", sprintf("(%.10f)", z), expr_str, fixed = TRUE)
-
-    tryCatch({
-      eval(parse(text = expr_str))
-    }, error = function(e) {
-      stop(sprintf(
-        "Failed to parse symmetry operation: '%s' -> '%s'",
-        op_str,
-        expr_str
-      ))
-    })
+    # Increase internal string precision to 16 to prevent truncation errors
+    # with fractional coordinates (e.g., 1/3) before evaluation.
+    expr_str <- gsub("x", sprintf("(%.16f)", x), op_str, fixed = TRUE)
+    expr_str <- gsub("y", sprintf("(%.16f)", y), expr_str, fixed = TRUE)
+    expr_str <- gsub("z", sprintf("(%.16f)", z), expr_str, fixed = TRUE)
+    tryCatch(
+      eval(parse(text = expr_str)),
+      error = function(e)
+        stop(paste("SymOp parse error:", op_str))
+    )
   }
 
   n_atoms <- nrow(atomic_coordinates)
   n_ops <- nrow(symmetry_operations)
 
-  # Expand structure using optimized lapply (Vectorized approach, no set() loops)
   expanded_list <- lapply(seq_len(n_atoms), function(i) {
     row <- atomic_coordinates[i, ]
     lbl <- row$Label
-    val_x <- row$x_a
-    val_y <- row$y_b
-    val_z <- row$z_c
+    val_x <- as.numeric(row$x_a)
+    val_y <- as.numeric(row$y_b)
+    val_z <- as.numeric(row$z_c)
 
-    # Apply all symmetry operations to this specific atom
     new_coords <- lapply(seq_len(n_ops), function(j) {
       op_x <- symmetry_operations$x[j]
       op_y <- symmetry_operations$y[j]
@@ -58,39 +53,27 @@ apply_symmetry_operations <- function(atomic_coordinates, symmetry_operations) {
       data.table(
         Label = paste(lbl, j, sep = "_"),
         SymOp = j,
+        # Apply modulo initially to handle operations like x+1/2
[…]
