### 2026-01-27 3a0bc14
Implement PBC-aware site merging and update symmetry logic
Added merge_sites_pbc() to deduplicate atomic sites within a distance tolerance using hierarchical clustering and periodic boundary conditions, mimicking pymatgen's merge_sites. Updated apply_symmetry_operations() to use distance-based merging instead of coordinate rounding, and refactored expand_transformed_coords() to remove precision-based deduplication. Updated documentation and NAMESPACE accordingly. Also adjusted minimum_distance() to remove hardcoded distance filtering, and increased the default tolerance for floating-point nois
[…]
