+++ packages/crystract/R/process_cif_files.R
@@ -73,7 +73,7 @@ read_cif_files <- function(file_paths) {
 #' @param calculate_bond_angles Logical. If `TRUE`, computes bond angles.
 #' @param perform_error_propagation Logical. If `TRUE`, calculates uncertainties.
 #'   Any missing error values in the CIF file are treated as zero.
-#' @param tolerance Numeric. The cutoff distance (default 1e-6) below which
+#' @param tolerance Numeric. The cutoff distance (default 1e-4) below which
 #'   distances are considered floating-point noise and ignored. Increase this
 #'   (e.g. to 1e-4) if "ghost" bonds appear in disordered structures.
 #' @param minimum_distance_delta Numeric. The relative tolerance parameter used
@@ -103,7 +103,7 @@ analyze_single_cif <- function(cif_content,
                                bonding_algorithms = c("minimum_distance"),
                                calculate_bond_angles = TRUE,
                                perform_error_propagation = TRUE,
-                               tolerance = 1e-6,
+                               tolerance = 1e-4,
                                minimum_distance_delta = 0.1) {
   # --- Step 0: Handle Input Type ---
   if (is.character(cif_content)) {
@@ -159,19 +159,18 @@ analyze_single_cif <- function(cif_content,
 
   # --- Step 4: Calculations, Transformations, and Expansions ---
   if (perform_calcs_and_transforms) {
-    # If tolerance is 1e-4, we use precision 3 (0.001) for deduplication.
-    # This prevents '0.33332' and '0.33337' (diff 5e-5) from being split into '0.3333' and '0.3334'.
-    if (tolerance <= 0) {
-      precision_val <- 15
-    } else {
-      # Subtract 1 to create a buffer
-      precision_val <- max(0, as.integer(ceiling(-log10(tolerance))) - 1)
-    }
+    # No precision calculation anymore. We use distance clustering.
 
-    transformed_coords <- apply_symmetry_operations(atomic_coordinates, symmetry_operations, precision = precision_val)
+    transformed_coords <- apply_symmetry_operations(
+      atomic_coordinates,
+      symmetry_operations,
+      unit_cell_metrics,
+      tolerance = tolerance # This is used for clustering
+    )
 
-    expanded_coords <- expand_transformed_coords(transformed_coords, precision = precision_val)
+    expanded_coords <- expand_transformed_coords(transformed_coords)
 
+    # tolerance used here for filtering floating point noise in distance calculation
     distances <- calculate_distances(atomic_coordinates,
                                      expanded_coords,
                                      unit_cell_metrics,
@@ -293,9 +292,7 @@ analyze_cif_files <- function(file_paths,
                               output_dir = NULL,
                               batch_size = 1000,
                               ...) {
-  # --- 1. Handle Input: File Paths or Pre-parsed List ---
   if (is.character(file_paths)) {
-    # Check if the input is a single directory path
     if (length(file_paths) == 1 && dir.exists(file_paths)) {
       message(sprintf(
         "Input is a directory. Searching for .cif files in '%s'...",
@@ -307,30 +304,25 @@ analyze_cif_files <- function(file_paths,
         full.names = TRUE,
         ignore.case = TRUE
       )
-
       if (length(found_files) == 0) {
         warning(sprintf("No .cif files found in directory '%s'.", file_paths))
-        # Namespaced data.table
         return(data.table::data.table())
       }
       file_paths <- found_files
     }
-
     cif_contents_list <- read_cif_files(file_paths)
   } else if (is.list(file_paths) &&
              (length(file_paths) == 0 ||
[…]
