+++ packages/crystract/R/process_coordinates.R
@@ -1,15 +1,86 @@
+#' @title Merge Close Atoms (Pymatgen Style)
+#' @description Merges atomic sites that are within a specified distance tolerance,
+#' accounting for Periodic Boundary Conditions (PBC). This mimics pymatgen's
+#' `Structure.merge_sites` logic using hierarchical clustering.
+#'
+#' @param atomic_coordinates A data.table with columns `Label`, `x_a`, `y_b`, `z_c`.
+#' @param unit_cell_metrics A data.table containing cell lengths and angles.
+#' @param tol Numeric. Distance tolerance in Angstroms. Sites closer than this will be merged.
+#' @return A deduplicated data.table with coordinates averaged.
+#' @export
+merge_sites_pbc <- function(atomic_coordinates,
+                            unit_cell_metrics,
+                            tol = 0.1) {
+  if (nrow(atomic_coordinates) < 2)
+    return(atomic_coordinates)
+
+  # 1. Extract Cell Parameters
+  a <- as.numeric(unit_cell_metrics$`_cell_length_a`)
+  b <- as.numeric(unit_cell_metrics$`_cell_length_b`)
+  c <- as.numeric(unit_cell_metrics$`_cell_length_c`)
+  alpha <- as.numeric(unit_cell_metrics$`_cell_angle_alpha`) * pi / 180
+  beta  <- as.numeric(unit_cell_metrics$`_cell_angle_beta`) * pi / 180
+  gamma <- as.numeric(unit_cell_metrics$`_cell_angle_gamma`) * pi / 180
+
+  # 2. Metric Tensor components
+  g11 <- a^2
+  g22 <- b^2
+  g33 <- c^2
+  g12 <- a * b * cos(gamma)
+  g13 <- a * c * cos(beta)
+  g23 <- b * c * cos(alpha)
+
+  # 3. Calculate Distance Matrix with PBC manually
+  coords <- as.matrix(atomic_coordinates[, .(x_a, y_b, z_c)])
+  n <- nrow(coords)
+  dist_mat <- matrix(0, nrow = n, ncol = n)
+
+  for (i in 1:(n - 1)) {
+    for (j in (i + 1):n) {
+      dx <- coords[i, 1] - coords[j, 1]
+      dy <- coords[i, 2] - coords[j, 2]
+      dz <- coords[i, 3] - coords[j, 3]
+
+      # Apply Minimum Image Convention
+      dx <- dx - round(dx)
+      dy <- dy - round(dy)
+      dz <- dz - round(dz)
+
+      d2 <- g11 * dx^2 + g22 * dy^2 + g33 * dz^2 +
+        2 * g12 * dx * dy + 2 * g13 * dx * dz + 2 * g23 * dy * dz
+
+      d <- sqrt(d2)
+      dist_mat[i, j] <- d
+      dist_mat[j, i] <- d
+    }
+  }
+
+  # 4. Hierarchical Clustering
+  dist_obj <- as.dist(dist_mat)
+  # Single linkage groups chain-linked neighbors (standard for merging)
+  hc <- hclust(dist_obj, method = "single")
+  clusters <- cutree(hc, h = tol)
+
+  # 5. Merge Sites (Take first entry per cluster to preserve labels)
+  merged_dt <- atomic_coordinates[, .SD[1], by = clusters]
+  merged_dt[, clusters := NULL]
+
+  return(merged_dt)
+}
+
 #' @title Apply Symmetry Operations to Generate a Full Unit Cell
 #' @description Generates all symmetry-equivalent atomic positions within the unit cell.
 #' @param atomic_coordinates A `data.table` of asymmetric atoms.
 #' @param symmetry_operations A `data.table` of operations.
[…]
