### 2026-01-27 36f73dc
Revise and expand crystract package description
Expanded the description of the 'crystract' R package to clarify its tested compatibility with major databases, detail supported neighbor-finding algorithms, and provide the GitHub repository link. Minor wording improvements and typo corrections were also made for clarity and accuracy.
