+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
   - name: Anirudh Prabhu
     orcid: 0000-0002-9921-6084
     affiliation: 1
-    corresponding: true 
+    corresponding: true
 affiliations:
  - name: Earth and Planets Laboratory, Carnegie Institution for Science, Washington, DC, USA
    index: 1
@@ -38,7 +38,7 @@ The Crystallographic Information File (CIF) is the standard format for dissemina
 
 Despite the standardization provided by the CIF format, significant practical barriers remain for researchers aiming to perform high-throughput computational analysis, particularly in batch. The first barrier is technical: CIF files, while standardized, often exhibit syntactic variations or errors depending on their originating software or laboratory, which can cause simplistic parsers to fail. The second barrier is conceptual: a CIF file does not typically contain an explicit list of all atoms in the unit cell. Instead, it reports the unique atoms in the asymmetric unit and a set of symmetry operations. A complete structural analysis, therefore, requires the correct application of these symmetry operations to generate the full unit cell and its atomic contents—a non-trivial, error-prone computational task that must be handled by specialized software. This complexity often forces researchers into a fragmented and inefficient workflow, piecing together disparate tools for data validation, structure generation, geometric analysis, and final statistical modeling.
 
-A number of excellent software tools have been developed to address some of these challenges, yet a comprehensive review reveals a specific and critical gap. The most mature ecosystem for computational materials science currently resides in Python, where `pymatgen` stands as a powerful and widely adopted standard [@Ong2013]. This extensive library offers robust CIF parsing and a host of advanced analysis functions, including the calculation of interatomic distances and the prediction of bonding environments using sophisticated, data-mining approaches like the CrystalNN algorithm [@Pan2021]. While `pymatgen` offers robust CIF parsing and advanced analysis, it has limitations for a complete, statistically rigorous workflow: it does not natively compute bond angles, nor does it programmatically propagate the experimental uncertainties reported in CIFs for its derived geometric quantities. Furthermore, batch processing requires the user to write custom scripts to loop over files. Filtering presents another challenge: although `pymatgen` includes a function to filter for minimum and maximum distances, they have to be specified one element at a time, one file at a time, by the user. This makes it cumbersome to use `pymatgen` for large scale datasets and studies.
+A number of excellent software tools have been developed to address some of these challenges, yet a comprehensive review reveals a specific and critical gap. The most mature ecosystem for computational materials science currently resides in Python, where `pymatgen` stands as a powerful and widely adopted standard [@Ong2013]. This extensive library offers robust CIF parsing and a host of advanced analysis functions, including the calculation of interatomic distances and the prediction of bonding environments using sophisticated, data-mining approaches like the CrystalNN algorithm [@Pan2021]. While `pymatgen` offers robust CIF parsing and advanced analysis, it has limitations for a complete, statistically rigorous workflow: it does not programmatically propagate the experimental uncertainties reported in CIFs for its derived geometric quantities. Furthermore, batch processing requires the user to write custom scripts to loop over files. Filtering presents another challenge: although `pymatgen` includes a function to filter for minimum and maximum distances, they have to be specified one element at a time, one file at a time, by the user. This makes it cumbersome to use `pymatgen` for large scale datasets and studies.
 
[…]
