### 2026-01-28 #9529 @anirudhprabhu
@zhubonan @singularitti @bobleesj @hisacro

Hello everyone,
My name is Anirudh Prabhu and I am Don’s mentor and the senior author on this paper. Thank you so much for your deep dive into the package and the manuscript to flesh out the need for more detail and for your suggestions for improvements in both its functionality and usability. I also appreciate your patience in working with Don on this review, since this will be his first scientific paper and thus this is his very first interaction with editors and reviewers. I have been following this and other review comment chains carefully, and noticed that while all of the tasks suggested by the reviewers are valid and will be part of our next steps, some of these tasks will take a longer time, including the release of this package to CRAN.
To put things in context, we are working with mineralogists and material scientists on a parallel scientific study using this package to analyze the behavior of Clathrate I type compound systems and Chalcogen minerals and ores. My intention was to perform user scenario testing and a real scientific advance made using crystract before releasing it on CRAN. That way both publications will be out by the time the package is released on CRAN. In the meantime, users can install the package on R and Rstudio using github installation function.
All this said, I would like this to be a widely used community resource for mineralogists, crystallographers and material scientists alike and take the comments left by the reviewers very seriously. I would also like to make this review process as smooth for you all, given the time you are dedicating to this paper.
So, I would like to pose this question to the reviewers and editors. If you think it would be more prudent to publish this package on CRAN before this paper is published, then we are happy to do so, but this will probably take quite a bit more time for us. If this is the case, do you suggest we pause or withdraw the review on this version of the paper and resubmit with a CRAN approved version of the package and with any of the other improvements to the code we will work on in the next few months?

Alternatively, we could keep this review going, but at some point it begins to feel like we are iteratively improving both our paper and package through much needed suggestions, but taking up the reviewer’s valuable time.

Thanks again for your help in shaping this manuscript and for working with Don.
Anirudh Prabhu
