### 2026-01-28 #9529 @singularitti
Hi @anirudhprabhu, I am fine with whatever options you prefer. But before that, could you briefly clarify your understanding of what is required to publish an R package on CRAN (e.g., the scope of CRAN checks and review, and whether this involves scientific peer review versus technical and policy compliance)? I am not a CRAN user, so I am unfamiliar with its procedures.
