### 2026-01-30 #9529 @zhubonan
Hi @anirudhprabhu and @Don-Tat I am not an expert on R. Is CRAN like PyPI? If so, I think having it released there is not required for your JOSS review.
Ideally, we would like to have the package polished and ready for review before submission.  But if you feel you more time to polish the software please let me and the reviewers know and give us a ping when you think it is ready again.
