### 2026-02-02 6b218cf
Add bonding algorithms and data enhancements
Introduces new and updated bonding algorithms (BrunnerNNReciprocal, EconNN, CrystalNN, VoronoiNN) and refactors bonding logic into predict_bond and s.R. Adds Shannon Ionic Radii data, Pauling electronegativity data and extraction utilities, updates atomic coordinate extraction to support oxidation states and thermal parameters, and improves algorithm selection in CIF analysis. Updates documentation and NAMESPACE to reflect new features and imports.
