+++ packages/crystract/R/calculate_properties.R
@@ -10,21 +10,6 @@
 #' @return A `data.table` of all non-zero distances.
 #' @family property calculators
 #' @export
-#' @examples
-#' # Example setup
-#' cif_file <- system.file("extdata", "ICSD422.cif", package = "crystract")
-#' if (file.exists(cif_file)) {
-#'   cif_content <- read_cif_files(cif_file)[[1]]
-#'   atoms <- extract_atomic_coordinates(cif_content)
-#'   metrics <- extract_unit_cell_metrics(cif_content)
-#'   sym_ops <- extract_symmetry_operations(cif_content)
-#'   full_cell <- apply_symmetry_operations(atoms, sym_ops)
-#'   super_cell <- expand_transformed_coords(full_cell)
-#'
-#'   # Increase tolerance if you see tiny "ghost" distances (e.g. 1e-5)
-#'   dists <- calculate_distances(atoms, super_cell, metrics, tolerance = 1e-4)
-#'   print(head(dists))
-#' }
 calculate_distances <- function(atomic_coordinates,
                                 expanded_coords,
                                 unit_cell_metrics,
@@ -80,19 +65,19 @@ calculate_distances <- function(atomic_coordinates,
   return(distances)
 }
 
-#' @title Calculate Neighbor Counts
+#' @title Calculate Coordination Numbers
 #' @description Counts the number of nearest neighbors for each central atom based on
 #'   a table of bonded pairs.
 #' @param bonded_pairs_table A `data.table` of bonded pairs.
-#' @return A `data.table` with columns 'Atom' and 'NeighborCount'.
+#' @return A `data.table` with columns 'Atom' and 'CoordinationNumber'.
 #' @family property calculators
 #' @export
 calculate_neighbor_counts <- function(bonded_pairs_table) {
   if (is.null(bonded_pairs_table) || nrow(bonded_pairs_table) == 0) {
-    return(data.table(Atom = character(), NeighborCount = integer()))
+    return(data.table(Atom = character(), CoordinationNumber = integer()))
   }
 
-  neighbor_counts <- bonded_pairs_table[, .(NeighborCount = .N), by = .(Atom1)]
+  neighbor_counts <- bonded_pairs_table[, .(CoordinationNumber = .N), by = .(Atom1)]
 
   setnames(neighbor_counts, "Atom1", "Atom")
 
@@ -547,175 +532,6 @@ propagate_angle_error <- function(bond_angles,
   ))
 }
 
-#' @title Identify Atomic Bonds using the Minimum Distance Method
-#' @description Identifies bonded atoms by finding the nearest neighbor distance (d_min)
-#'   for each central atom and defining a cutoff distance (d_cut) as
-#'   d_cut = (1 + delta) * d_min.
-#' @param distances A `data.table` of interatomic distances from `calculate_distances`.
-#' @param delta The relative tolerance parameter (default 0.1 or ten percent).
-#' @return A `data.table` of bonded pairs.
-#' @family bonding algorithms
-#' @export
-#' @examples
-#' cif_file <- system.file("extdata", "ICSD422.cif", package = "crystract")
-#' if (file.exists(cif_file)) {
-#'   # Setup code
-#'   cif_content <- read_cif_files(cif_file)[[1]]
-#'   atoms <- extract_atomic_coordinates(cif_content)
-#'   metrics <- extract_unit_cell_metrics(cif_content)
-#'   sym_ops <- extract_symmetry_operations(cif_content)
-#'   full_cell <- apply_symmetry_operations(atoms, sym_ops)
-#'   super_cell <- expand_transformed_coords(full_cell)
[…]
