+++ packages/crystract/R/crystract-package.R
@@ -5,7 +5,7 @@
 #' Files (.cif), extracting essential data such as chemical formulas, unit cell
 #' parameters, atomic coordinates, and symmetry operations. It also includes
 #' tools to calculate interatomic distances, identify bonded pairs using various
-#' algorithms (Minimum Distance, Brunner's, Hoppe's), determine nearest
+#' algorithms (Minimum Distance, Brunner's, EconNN, CrystalNN, Voronoi), determine nearest
 #' neighbor counts, and calculate bond angles. The package is designed to
 #' facilitate the preparation of crystallographic data for further analysis,
 #' including machine learning applications in materials science.
@@ -16,6 +16,7 @@
 #' @importFrom data.table .N := .SD as.data.table copy data.table fread rbindlist set setkey setnames
 #' @importFrom dplyr distinct
 #' @importFrom stats as.dist cutree hclust na.omit
+#' @importFrom geometry delaunayn
 #' @importFrom stringr str_match str_split_fixed
 #' @importFrom utils combn globalVariables
 NULL
