### 2026-02-02 6f9a1cd
Enhance bonding algorithms and radius logic in crystract
Improves the EconNN, VoronoiNN, and CrystalNN bonding algorithms with better handling of oxidation states, cation-anion filtering, and Voronoi cell pathologies. Refactors get_radius to match pymatgen logic, including linear interpolation for fractional oxidation states and improved fallbacks. Adds a tetrahedron volume utility and semicircle integral for geometric calculations. Updates documentation to reflect new parameters and logic.
