+++ packages/crystract/R/utils_data.R
@@ -14,7 +14,7 @@ load_ext_data <- function(filename) {
 }
 
 #' @title Get Electronegativity
-#' @description Looks up Pauling electronegativity. matches pymatgen `Element.X`.
+#' @description Looks up Pauling electronegativity. Matches pymatgen `Element.X`.
 #' @param symbols Character vector of element symbols.
 #' @return Numeric vector of electronegativities.
 #' @noRd
@@ -31,68 +31,103 @@ get_electronegativity <- function(symbols) {
 }
 
 #' @title Get Atomic/Ionic Radius
-#' @description mimic pymatgen `_get_radius` logic.
+#' @description Mimics pymatgen `_get_radius` logic.
 #' 1. If oxidation state is provided and non-zero: Look up Shannon Radius.
-#' 2. If Shannon radius not found for that CN, try to average or find closest.
-#' 3. If oxidation state is 0 or NA: Look up Covalent/Atomic Radius.
+#' 2. If fractional oxidation state: Linear interpolation between floor/ceil states.
+#' 3. If no specific match: Use average cationic/anionic radius for that element.
+#' 4. Fallback (or if OS is 0/NA): Use Covalent/Atomic Radius.
 #'
 #' @param symbol Character. Element symbol.
 #' @param oxidation_state Numeric. Formal charge.
-#' @param cn Numeric. Coordination number (optional).
+#' @param cn Numeric. Coordination number (optional, default 6).
 #' @return Numeric radius in Angstroms.
 #' @noRd
 get_radius <- function(symbol,
                        oxidation_state = NA,
                        cn = 6) {
-  # Atomic/Covalent (Fallback)
+  # 1. Load Covalent/Atomic Radii (Ultimate Fallback)
   atomic_rads <- get_radii_data()
 
-  # Shannon (Ionic) - Load from the provided CSV
+  # Helper for default radius
+  get_default_radius <- function(sym) {
+    r_val <- atomic_rads[Symbol == sym & Type == "covalent"]$Radius
+    if (length(r_val) > 0)
+      return(r_val[1])
+    return(0)
+  }
+
+  # 2. Handle 0 or NA Oxidation State -> Return Default immediately
+  if (is.na(oxidation_state) || oxidation_state == 0) {
+    return(get_default_radius(symbol))
+  }
+
+  # 3. Load Shannon Radii
   shannon_rads <- load_ext_data("Shannon_Radii.csv")
-  if (!is.null(shannon_rads)) {
-    # Rename columns to match internal function expectations
-    data.table::setnames(
-      shannon_rads,
-      old = c("Element", "Charge", "Ionic Radius"),
-      new = c("Symbol", "OxidationState", "Radius"),
-      skip_absent = TRUE
-    )
+  if (is.null(shannon_rads)) {
+    return(get_default_radius(symbol))
   }
 
+  # Normalize columns
+  data.table::setnames(
+    shannon_rads,
+    old = c("Element", "Charge", "Ionic Radius"),
+    new = c("Symbol", "OxidationState", "Radius"),
+    skip_absent = TRUE
+  )
 
-  radius <- 1.0 # Ultimate fallback
-
-  if (!is.na(oxidation_state) &&
-      oxidation_state != 0 && !is.null(shannon_rads)) {
+  # Helper to fetch radius for a specific integer/float OS
+  fetch_radius_for_os <- function(sym, os) {
     # Try exact match for Symbol + OS + CN
-    match <- shannon_rads[Symbol == symbol &
-                            OxidationState == oxidation_state &
+    match <- shannon_rads[Symbol == sym &
+                            OxidationState == os &
[…]
