+++ packages/crystract/R/utils_math.R
@@ -8,6 +8,22 @@ cross_product <- function(v1, v2) {
   c(v1[2] * v2[3] - v1[3] * v2[2], v1[3] * v2[1] - v1[1] * v2[3], v1[1] * v2[2] - v1[2] * v2[1])
 }
 
+#' @title Calculate Tetrahedron Volume (Internal)
+#' @description Calculates the volume of a tetrahedron given 4 vertices.
+#' Used for reconstruction of Voronoi face areas.
+#' Matches pymatgen `vol_tetra`.
+#' @param vt1,vt2,vt3,vt4 Numeric vectors of length 3 (vertices).
+#' @return Numeric volume.
+#' @noRd
+vol_tetra <- function(vt1, vt2, vt3, vt4) {
+  # Formula: |dot((v1 - v4), cross((v2 - v4), (v3 - v4)))| / 6
+  diff1 <- vt1 - vt4
+  diff2 <- vt2 - vt4
+  diff3 <- vt3 - vt4
+  cp <- cross_product(diff2, diff3)
+  abs(sum(diff1 * cp)) / 6
+}
+
 #' @title Calculate Circumcenter (Internal)
 #' @description Calculates the circumcenter of a tetrahedron defined by 4 points.
 #' @param pts A 4x3 matrix of coordinates.
@@ -40,15 +56,15 @@ calculate_circumcenter <- function(pts) {
 
 #' @title Calculate Solid Angle (Internal)
 #' @description Calculates the solid angle subtended by a polygon face from a center point,
-#' following the method used in pymatgen (Oosterom & Strackee).
+#' following the Oosterom & Strackee method used in pymatgen `solid_angle`.
 #' @param center Numeric vector of length 3 (center atom coords).
 #' @param face_coords Matrix of face vertices (Nx3).
 #' @return Numeric solid angle in steradians.
 #' @noRd
 calculate_solid_angle <- function(center, face_coords) {
-  # Calculate displacement vectors from center to vertices
-  r <- t(t(face_coords) - center)
-  r_norms <- sqrt(rowSums(r^2))
+  # Compute displacement from center
+  disp <- t(t(face_coords) - center)
+  r_norm <- sqrt(rowSums(disp^2))
 
   angle <- 0
   n <- nrow(face_coords)
@@ -56,33 +72,39 @@ calculate_solid_angle <- function(center, face_coords) {
   if (n < 3)
     return(0)
 
-  # Iterate through the triangle fan (0, i, i+1 logic)
-  # Pymatgen implementation iterates 0, 1, 2... we use 1-based indexing
-  for (i in 2:(n - 1)) {
-    r0 <- r[1, ]
-    ri <- r[i, ]
-    rj <- r[i + 1, ]
+  # Iterate through the triangle fan (1, i, i+1 logic)
+  # Pymatgen logic: "for ii in range(1, len(disp) - 1): jj = ii + 1"
+  # This corresponds to indices 0, ii, jj in 0-indexed Python arrays.
+  # In R (1-based), this is indices 1, ii, jj where ii runs 2 to n-1.
 
-    n0 <- r_norms[1]
-    ni <- r_norms[i]
-    nj <- r_norms[i + 1]
+  for (ii in 2:(n - 1)) {
+    jj <- ii + 1
+
+    r0 <- disp[1, ] # Python index 0
+    ri <- disp[ii, ]
+    rj <- disp[jj, ]
+
+    n0 <- r_norm[1]
+    ni <- r_norm[ii]
+    nj <- r_norm[jj]
 
     # Scalar triple product: r0 . (ri x rj)
     cp <- cross_product(ri, rj)
     tp <- abs(sum(r0 * cp))
 
[…]
