+++ packages/crystract/man/voronoi_nn.Rd
@@ -9,7 +9,8 @@ voronoi_nn(
   expanded_coords,
   unit_cell_metrics,
   cutoff = 13,
-  tol = 0
+  tol = 0,
+  allow_pathological = FALSE
 )
 }
 \arguments{
@@ -22,13 +23,15 @@ voronoi_nn(
 \item{cutoff}{Distance cutoff (default 13.0).}
 
 \item{tol}{Tolerance for solid angle weights (default 0).}
+
+\item{allow_pathological}{Logical. If FALSE, raises error for infinite vertices.}
 }
 \value{
 A \code{data.table} of bonded pairs.
 }
 \description{
 Performs 3D Voronoi analysis on the supercell.
-strictly mapping Asymmetric Unit atoms to their Supercell equivalents for processing.
+strictly mapping Asymmetric Unit atoms to their Supercell equivalents.
 }
 \seealso{
 Other bonding algorithms: 

[…]
