### 2026-02-02 #9529 @anirudhprabhu
Hi @zhubonan and @singularitti
[CRAN](https://cran.r-project.org/web/packages/policies.html) is a quite a bit more restrictive than PyPI. It focuses on very specific aspects of package testing, including reverse dependencies etc. But the function it serves is the same as what PyPI in Python.

I did value comments left by @bobleesj on ease of package installation without the need for a C++ compiler, so wanted to take his suggestion to heart and see if we can enable that through CRAN submission or through some other means. The reason for the pause is that CRAN usually takes at least 3 to 4 weeks of review before release (depending on the number of submitted packages in review). I will keep you all updated on our CRAN submission process and we will update you once it is done.

I will let @Don-Tat continue communicating with you if you have any questions. I just wanted to step in provide some context for the decision to not submit to CRAN before submission.
