### 2026-02-18 4a2afee
Add chalcogen notebook; refactor crystract
Add a new RMarkdown analysis notebook (and generated HTML) for targeted chemistry analysis of sulfides/selenides/tellurides with diagnostics and visualization. Refactor crystract internals: update NAMESPACE, modify extract_cif_data.R, predict_bonds.R, process_cif_files.R, process_coordinates.R and utils_global.R; add utils_internal.R and remove legacy utils_data.R and utils_math.R. Update many man pages and add calculate_expansion_factors.Rd to reflect the API and documentation changes. Overall: consolidate utility code and improve processing/diagnost
[…]
