+++ packages/crystract/R/extract_cif_data.R
@@ -1,5 +1,4 @@
-#' Generic Value Extractor (Internal)
-#'
+#' @title Generic Value Extractor (Internal)
 #' @description Extracts a single value from CIF content based on a matching text pattern.
 #' @param cif_content A data.table containing the lines of a CIF file.
 #' @param pattern The text pattern (e.g., "_database_code_") to search for.
@@ -7,7 +6,6 @@
 #' @return A character string of the cleaned value, or NA if not found.
 #' @noRd
 extract_value <- function(cif_content, pattern, remove_pattern = TRUE) {
-  # fixed = TRUE automatically handles indented lines as it searches for the substring anywhere
   matching_lines <- cif_content[grepl(pattern, V1, fixed = TRUE)]
   if (nrow(matching_lines) > 0) {
     value <- matching_lines$V1[1]
@@ -105,7 +103,6 @@ extract_unit_cell_metrics <- function(cif_content) {
   }
 
   for (param in cell_parameters) {
-    # fixed = TRUE handles indented lines automatically
     line <- cif_content[grepl(param, V1, fixed = TRUE)]$V1
     if (length(line) > 0) {
       match <- stringr::str_match(line[1], "\\s+([0-9\\.]+)(?:\\(([0-9]+)\\))?")
@@ -184,27 +181,23 @@ extract_atom_type_info <- function(cif_content) {
 
   data_lines <- cif_content$V1[first_data_line:last_data_line]
 
-  # Parse using fread for robustness
   dt_raw <- data.table::fread(
     text = paste(data_lines, collapse = "\n"),
     header = FALSE,
     sep = "auto",
-    quote = ""
+    quote = "",
+    fill = TRUE
   )
 
-  # Helper to clean strings
   clean_val <- function(x)
     gsub("^'|'$|^\"|\"$", "", as.character(x))
 
-  # Extract Symbols
   if (sym_idx > ncol(dt_raw))
     return(NULL)
   symbols <- clean_val(dt_raw[[sym_idx]])
 
-  # Extract Oxidation (if present)
   oxi_states <- rep(NA_real_, nrow(dt_raw))
   if (length(ox_idx) > 0 && ox_idx <= ncol(dt_raw)) {
-    # Remove brackets for uncertainty e.g. "2.0(1)" -> 2.0
     raw_ox <- clean_val(dt_raw[[ox_idx]])
     raw_ox <- gsub("\\(.*\\)", "", raw_ox)
     oxi_states <- as.numeric(raw_ox)
@@ -215,9 +208,8 @@ extract_atom_type_info <- function(cif_content) {
 
 #' @title Extract Atomic Coordinates
 #' @description Parses atomic site info, including labels, fractional coordinates,
-#'   occupancies, oxidation states, and thermal parameters. It efficiently
-#'   normalizes atom labels to a 'SymbolNumber' format and validates them
-#'   against the chemical formula.
+#'   occupancies, oxidation states, and thermal parameters. It includes logic
+#'   to extract oxidation states from both site tags and type loops.
 #' @param cif_content A `data.table` containing the lines of a CIF file.
 #' @param chemical_formula The chemical formula string for validation.
 #' @return A `data.table` with atomic coordinate data, or `NULL` if not found.
@@ -225,13 +217,9 @@ extract_atom_type_info <- function(cif_content) {
 #' @export
 extract_atomic_coordinates <- function(cif_content, chemical_formula = NA) {
   # --- 0. Pre-fetch Atom Type Info (for Oxidation State Lookup) ---
-  # This part assumes a helper function extract_atom_type_info exists.
[…]
