+++ packages/crystract/R/process_cif_files.R
@@ -56,48 +56,22 @@ read_cif_files <- function(file_paths) {
 
 #' @title Analyze the Content of a Single CIF File
 #' @description The core worker function that orchestrates the analysis pipeline
-#'   for a single crystal structure's data. It is called by `analyze_cif_files`
-#'   for batch processing, but can now be called directly with a file path.
+#'   for a single crystal structure's data. It dynamically adjusts supercell
+#'   size based on the requested bonding algorithms to ensure Voronoi accuracy.
 #' @param cif_content Either a `data.table` containing the lines of a CIF file,
 #'   OR a character string specifying the file path to a CIF file.
 #' @param file_name The name of the original CIF file, used for labeling output.
-#'   If `NULL` (default) and a file path is provided for `cif_content`,
-#'   the filename is automatically extracted from the path.
-#' @param perform_extraction Logical. If `TRUE`, extracts all metadata and basic
-#'   structural data.
-#' @param perform_calcs_and_transforms Logical. If `TRUE`, generates the full
-#'   unit cell, expands it, and calculates interatomic distances.
-#' @param bonding_algorithms A character vector of bonding algorithms to use.
-#'   The first algorithm is used for subsequent angle/neighbor calculations.
-#'   Options: `"minimum_distance"`, `"brunner"`, `"econ"`, `"crystal_nn"`, `"voronoi"`.
-#'   Use `"none"` to skip.
-#' @param calculate_bond_angles Logical. If `TRUE`, computes bond angles for
-#'   each applied bonding algorithm.
-#' @param perform_error_propagation Logical. If `TRUE`, calculates uncertainties.
-#'   Any missing error values in the CIF file are treated as zero.
-#' @param tolerance Numeric. The cutoff distance (default 1e-4) below which
-#'   distances are considered floating-point noise and ignored. Increase this
-#'   (e.g. to 1e-4) if "ghost" bonds appear in disordered structures.
-#' @param minimum_distance_delta Numeric. The relative tolerance parameter used
-#'   specifically for the "minimum_distance" bonding algorithm. Default is 0.1.
-#'   (e.g., d_cut = d_min * (1 + delta)).
-#' @return A one-row `data.table` with results. If essential data for calculations
-#'   is missing, a warning is issued and only partial results are returned.
+#' @param perform_extraction Logical. If `TRUE`, extracts all metadata.
+#' @param perform_calcs_and_transforms Logical. If `TRUE`, generates unit cell
+#'   and supercell.
+#' @param bonding_algorithms Character vector. Options: `"minimum_distance"`,
+#'   `"brunner"`, `"econ"`, `"crystal_nn"`, `"voronoi"`.
+#' @param calculate_bond_angles Logical.
+#' @param perform_error_propagation Logical.
+#' @param tolerance Numeric. Cutoff for floating-point noise and atom merging (default 1e-4).
+#' @param minimum_distance_delta Numeric. Tolerance for min-dist algorithm.
+#' @return A one-row `data.table` with results.
 #' @export
-#' @examples
-#' cif_file <- system.file("extdata", "ICSD422.cif", package = "crystract")
-#' if (file.exists(cif_file)) {
-#'   # Option 1: Pass the file path directly
-#'   result_from_path <- analyze_single_cif(cif_file)
-#'
-#'   # Option 2: Pass pre-loaded content with custom bonding delta
-#'   cif_content <- read_cif_files(cif_file)[[1]]
-#'   result_from_dt <- analyze_single_cif(
-#'     cif_content,
-#'     basename(cif_file),
-#'     minimum_distance_delta = 0.15
-#'   )
-#' }
 analyze_single_cif <- function(cif_content,
                                file_name = NULL,
                                perform_extraction = TRUE,
[…]
