### 2026-02-18 920e58f
Refactor crystract pkgdown config
Normalize and reorganize packages/crystract/_pkgdown.yml: clean YAML formatting and quoting, reformat navbar structure, and adjust GitHub aria-label. Update the reference sections by adding/renaming bonding and coordinate helpers (brunner_nn_reciprocal, econ_nn, voronoi_nn, crystal_nn, calculate_expansion_factors, merge_sites_pbc) and ensure core, calculators, error propagation, filtering, data extraction, and Data entries list key functions. Remove commented lines and tidy spacing.
