### 2026-02-19 87ca387
Update pkgdown.yaml, DESCRIPTION title rewritten
