+++ packages/crystract/.github/workflows/pkgdown.yaml
@@ -37,7 +37,13 @@ jobs:
           needs: website
 
       - name: Build site
-        run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
+        run: |
+          pkgdown::build_site_github_pages(
+            pkg = "packages/crystract",
+            dest_dir = "docs",
+            new_process = FALSE,
+            install = FALSE
+          )
         shell: Rscript {0}
 
       - name: Deploy to GitHub pages 🚀
