### 2026-02-19 4749d0b
Fix pkgdown workflow working-dir and paths
