### 2026-02-19 d6561be
Simplify crystract pkgdown workflow and docs path
