+++ packages/crystract/_pkgdown.yml
@@ -75,4 +75,3 @@ reference:
   contents:
   - covalent_radii
   - set_radii_data
-
