### 2026-02-19 4cab060
Create Jekyll workflow for GitHub Pages deployment
