+++ reference/crystract-package.md
@@ -1,0 +1,17 @@
+# crystract: Tools to Extract Data from CIF Files for Crystallography
+
+Provides a suite of functions to parse Crystallographic Information
+Files (.cif), extracting essential data such as chemical formulas, unit
+cell parameters, atomic coordinates, and symmetry operations. It also
+includes tools to calculate interatomic distances, identify bonded pairs
+using various algorithms (Minimum Distance, Brunner's, EconNN,
+CrystalNN, Voronoi), determine nearest neighbor counts, and calculate
+bond angles. The package is designed to facilitate the preparation of
+crystallographic data for further analysis, including machine learning
+applications in materials science.
+
+## See also
+
+Useful links:
+
+- <https://prabhulab.github.io/ml-crystals/>
