+++ authors.md
@@ -1,0 +1,19 @@
+# Authors and Citation
+
+## Authors
+
+- **Don Ngo**. Maintainer.
+
+## Citation
+
+Ngo D (2026). *crystract: Crystallographic Information File (CIF) Data
+Processing Tools*. R package version 1.0.1,
+<https://prabhulab.github.io/ml-crystals/>.
+
+    @Manual{,
+      title = {crystract: Crystallographic Information File (CIF) Data Processing Tools},
+      author = {Don Ngo},
+      year = {2026},
+      note = {R package version 1.0.1},
+      url = {https://prabhulab.github.io/ml-crystals/},
+    }
