### 2026-02-19 #9529 @Don-Tat
Hi @zhubonan,
Thanks for checking in. No updates on the CRAN submission yet, but we will let you know as soon as we have something to share on this. In the meantime, I am actively working to address some of the other comments from reviewers and we will keep pushing on the non-CRAN suggestions from reviewers.
