### 2026-02-23 9c78cc8
Modified CrystalNN weighting and internal data
Refactor CrystalNN neighbor/bond prediction and add packaged internal data. crystal_nn was rewritten to clarify behavior, add options (cation_anion, weighted_cn), rename porosity_adjustment -> porous_adjustment, use internal electronegativity/radii lookups, apply smooth distance cutoffs, and compute final neighbor weights via a semicircle-integral based probability scheme. utils_internal.R now exposes preloaded shannon_radii and pauling_en lookups (removing repeated disk I/O) and fixes a small solid-angle branch; man page for crystal_nn was update
[…]
