+++ packages/crystract/R/predict_bonds.R
@@ -361,18 +361,28 @@ voronoi_nn <- function(atomic_coordinates,
 }
 
 #' @title Identify Atomic Bonds using CrystalNN
-#' @description Port of Pymatgen's `CrystalNN` weighting logic. Includes Porosity,
-#' Electronegativity, and Distance penalties.
-#' Uses specific Decision Tree logic for radii (Shannon -> Covalent -> Atomic) per
-#' pymatgen's fallback logic.
-#' @param distances Ignored.
+#' @description Poer of Pymatgen's `CrystalNN` algorithm from the ground up.
+#' It uses a Voronoi-based algorithm and solid angle weights to determine the
+#' probability of various coordination environments. It modifies probability using
+#' smooth distance cutoffs and Pauling electronegativity differences. The output
+#' is either the most probable coordination environment (`weighted_cn = FALSE`)
+#' or a weighted list of coordination environments (`weighted_cn = TRUE`).
+#' @param distances Ignored (CrystalNN generates its own Voronoi basis).
 #' @param atomic_coordinates Primary atom set.
 #' @param expanded_coords Expanded supercell.
 #' @param unit_cell_metrics Cell parameters.
-#' @param cutoff_length Voronoi cutoff. Default 7.0 matches Pymatgen CrystalNN default.
-#' @param x_diff_weight Electronegativity weight (default 3.0).
-#' @param porosity_adjustment Logical (default TRUE).
-#' @param distance_cutoffs Vector c(0.5, 1.0).
+#' @param cutoff_length Numeric. Cutoff in Angstroms for initial neighbor search
+#' (default 7.0).
+#' @param x_diff_weight Numeric. Electronegativity difference weight (default 3.0).
+#' @param porous_adjustment Logical. If TRUE, adjusts Voronoi weights to
+#' better describe layered / porous structures (default TRUE).
+#' @param distance_cutoffs Numeric vector. Penalizes neighbor distances greater than
+#' sum of covalent radii plus these cutoffs. Set to NULL to disable.
+#' @param cation_anion Logical. If TRUE, restricts bonding targets to sites
+#' with opposite or zero charge (requires oxidation states).
+#' @param weighted_cn Logical. If FALSE (default), returns neighbors for the
+#' most probable coordination environment with weight 1.0. If TRUE, returns all
+#' potential neighbors with fractional probabilities.
 #' @return A `data.table` of bonded pairs.
 #' @family bonding algorithms
 #' @export
@@ -381,10 +391,12 @@ crystal_nn <- function(distances,
                        expanded_coords,
                        unit_cell_metrics,
                        cutoff_length = 7.0,
-                       # Pymatgen default for CrystalNN is 7.0 (search_cutoff)
                        x_diff_weight = 3.0,
-                       porosity_adjustment = TRUE,
-                       distance_cutoffs = c(0.5, 1.0)) {
+                       porous_adjustment = TRUE,
+                       distance_cutoffs = c(0.5, 1.0),
+                       cation_anion = FALSE,
+                       weighted_cn = FALSE) {
+
   # 1. Get Base Voronoi Neighbors
   base <- voronoi_nn(
     atomic_coordinates,
@@ -410,22 +422,29 @@ crystal_nn <- function(distances,
[…]
