+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   clathrates: clathrates.html
   crystract: crystract.html
-last_built: 2026-02-19T17:46Z
+last_built: 2026-02-23T21:09Z
 urls:
   reference: https://prabhulab.github.io/ml-crystals/reference
   article: https://prabhulab.github.io/ml-crystals/articles
