### 2026-02-24 0cfd9b2
Refactor natural analysis and improve bond logic
Major revamp of the naturalTest RMarkdown and small API/robustness fixes in crystract bonding code. In the notebook: reordered outputs, added plotly for interactive visualizations, introduced a custom radii table, and implemented a robust per-structure worker (process_structure_robust) with batched parallel processing (future_lapply) that saves batch RDSs and logs filter stats. Aggregation/diagnosis logic was streamlined (high-CN debugging, weighted CN calculation and merging), and many plotting sections were converted to interactive ggplotly pl
[…]
