+++ packages/crystract/R/predict_bonds.R
@@ -62,13 +62,10 @@ brunner_nn_reciprocal <- function(distances, tol = 0.0) {
 
 #' @title Identify Atomic Bonds using Hoppe's EconNN Method
 #' @description Uses Effective Coordination Numbers (ECoN) and Mean Fictive Ionic Radii (MEFIR).
-#' Matches pymatgen logic regarding MEFIR initialization and iteration.
 #' @param distances A `data.table` of interatomic distances.
 #' @param atomic_coordinates A `data.table` of atomic coordinates used to map species to radii.
 #' @param tol A bond strength cutoff (default 0.2).
-#' @param use_fictive_radius Logical. If `TRUE`, calculates Hoppe's fictive ionic radius based
-#' on the ratio of central/neighbor radii. If `FALSE` (default), uses the raw bond distance,
-#' matching pymatgen's default behavior.
+#' @param use_fictive_radius Logical. If `TRUE`, calculates Hoppe's fictive ionic radius.
 #' @return A `data.table` of bonded pairs.
 #' @family bonding algorithms
 #' @export
@@ -82,14 +79,13 @@ econ_nn <- function(distances,
   dt <- copy(distances)
 
   if (use_fictive_radius) {
-    # Check if OxidationState exists, else create it as numeric NA
     atom_info <- copy(atomic_coordinates)
     if (!"OxidationState" %in% names(atom_info))
       atom_info[, OxidationState := NA_real_]
 
-    # Helper to vectorise get_radius_pymatgen_style
-    get_r_vec <- Vectorize(get_radius_decision_tree,
-                           vectorize.args = c("symbol", "oxidation_state"))
+    get_r_ionic <- Vectorize(get_ionic_radius,
+                             vectorize.args = c("symbol", "oxidation_state"))
+    get_r_def <- Vectorize(get_default_radius, vectorize.args = "symbol")
 
     dt[, `:=`(
       Sym1 = sub("[0-9].*", "", sub("_.*", "", Atom1)),
@@ -98,33 +94,30 @@ econ_nn <- function(distances,
       Parent2 = sub("_.*", "", Atom2)
     )]
 
-    # Merge Oxidation States
     dt[atom_info, on = c("Parent1" = "Label"), OS1 := i.OxidationState]
     dt[atom_info, on = c("Parent2" = "Label"), OS2 := i.OxidationState]
 
-    # Get Radii using standard Pymatgen logic
-    dt[, R1 := get_r_vec(Sym1, OS1, cn = 6)]
-    dt[, R2 := get_r_vec(Sym2, OS2, cn = 6)]
+    dt[, R1 := get_r_ionic(Sym1, OS1, cn = 6)]
+    dt[, R2 := get_r_ionic(Sym2, OS2, cn = 6)]
 
-    # Fictive Radii (Hoppe Eq 1)
+    # Fallbacks for fictive radius math
+    dt[R1 == 0, R1 := get_r_def(Sym1)]
+    dt[R2 == 0, R2 := get_r_def(Sym2)]
+
     dt[, RadiusSum := R1 + R2]
     dt[, FictiveR := ifelse(RadiusSum > 0, Distance * (R1 / RadiusSum), 0)]
   } else {
-    # If not using fictive radius, MEFIR is calculated on raw distances
     dt[, FictiveR := Distance]
   }
 
-  # Iterative MEFIR Solver (matches pymatgen _get_mean_fictive_ionic_radius)
   calc_mefir_group <- function(fictive_r) {
     min_v <- min(fictive_r, na.rm = TRUE)
     if (min_v <= 0 || !is.finite(min_v))
       return(0)
 
     mean_fir <- min_v
-
     for (i in 1:100) {
[…]
