+++ reference/econ_nn.md
@@ -1,8 +1,7 @@
 # Identify Atomic Bonds using Hoppe's EconNN Method
 
 Uses Effective Coordination Numbers (ECoN) and Mean Fictive Ionic Radii
-(MEFIR). Matches pymatgen logic regarding MEFIR initialization and
-iteration.
+(MEFIR).
 
 ## Usage
 
@@ -26,9 +25,7 @@ econ_nn(distances, atomic_coordinates, tol = 0.2, use_fictive_radius = FALSE)
 
 - use_fictive_radius:
 
-  Logical. If `TRUE`, calculates Hoppe's fictive ionic radius based on
-  the ratio of central/neighbor radii. If `FALSE` (default), uses the
-  raw bond distance, matching pymatgen's default behavior.
+  Logical. If `TRUE`, calculates Hoppe's fictive ionic radius.
 
 ## Value
 
