+++ reference/crystal_nn.md
@@ -1,12 +1,6 @@
 # Identify Atomic Bonds using CrystalNN
 
-Poer of Pymatgen's `CrystalNN` algorithm from the ground up. It uses a
-Voronoi-based algorithm and solid angle weights to determine the
-probability of various coordination environments. It modifies
-probability using smooth distance cutoffs and Pauling electronegativity
-differences. The output is either the most probable coordination
-environment (`weighted_cn = FALSE`) or a weighted list of coordination
-environments (`weighted_cn = TRUE`).
+Rebuild of Pymatgen's `CrystalNN` algorithm from the ground up.
 
 ## Usage
 
@@ -45,33 +39,28 @@ crystal_nn(
 
 - cutoff_length:
 
-  Numeric. Cutoff in Angstroms for initial neighbor search (default
-  7.0).
+  Numeric. Cutoff in Angstroms for initial neighbor search.
 
 - x_diff_weight:
 
-  Numeric. Electronegativity difference weight (default 3.0).
+  Numeric. Electronegativity difference weight.
 
 - porous_adjustment:
 
-  Logical. If TRUE, adjusts Voronoi weights to better describe layered /
-  porous structures (default TRUE).
+  Logical. If TRUE, adjusts Voronoi weights.
 
 - distance_cutoffs:
 
   Numeric vector. Penalizes neighbor distances greater than sum of
-  covalent radii plus these cutoffs. Set to NULL to disable.
+  radii.
 
 - cation_anion:
 
-  Logical. If TRUE, restricts bonding targets to sites with opposite or
-  zero charge (requires oxidation states).
+  Logical. Restrictions targets to opposite charge.
 
 - weighted_cn:
 
-  Logical. If FALSE (default), returns neighbors for the most probable
-  coordination environment with weight 1.0. If TRUE, returns all
-  potential neighbors with fractional probabilities.
+  Logical. Return fractional probabilities vs strict max probability.
 
 ## Value
 
