+++ reference/voronoi_nn.md
@@ -1,11 +1,6 @@
 # Identify Atomic Bonds using Voronoi Tessellation
 
-Performs 3D Voronoi analysis on the supercell. strictly mapping
-Asymmetric Unit atoms to their Supercell equivalents for processing.
-Ensures potential neighbors are within cutoff to avoid edge artifacts.
-Uses QJ (Joggled input) to handle degenerate structures (FCC/BCC) and
-filters/merges the resulting micro-faces to produce accurate Solid
-Angles.
+Performs 3D Voronoi analysis on the supercell.
 
 ## Usage
 
