### 2026-02-27 d7f9669
Add weighted CN, snapping, and NN/radii improvements
Introduce calculate_weighted_neighbor_counts (handles fractional occupancies, validates sums >1 and returns NULL, optimizes full-occupancy case) and export it. Integrate weighted CN calculation into CIF processing. Add snap_to_fraction and use it across extract/process routines to snap/wrap coordinates and occupancies to common crystallographic fractions. Refactor CrystalNN: clarify defaults, apply porosity/electronegativity/distance penalty logic, round weights, and simplify output; implement a radii decision-tree (Shannon -> covalent -> at
[…]
