### 2026-02-27 44e16d7
Add file-level weighted CN summary table
Add a new Section 5.2 to notebook/natural/naturalTest.Rmd that produces a file-by-file summary of per-site coordination numbers. The R chunk aggregates final_bonds_data into File, Site, Element, Geometric CN and Weighted CN (rounded to 3 decimals), sorts by file and site, and renders an interactive datatable with top filters and export buttons. If final_bonds_data is missing or empty, the chunk prints a fallback message.
