### 2026-02-27 88a6521
Revise crystract vignette examples and workflow
Update the crystract vignette to use runnable code chunks and modernized examples: change chunk eval flags to TRUE, add library(data.table), and update function names/column references (e.g., bonded_pairs_minimum_distance -> bonds_minimum_distance). Expand and standardize bonding algorithm examples (minimum_distance, brunner, econ, voronoi, crystal_nn), add clearer notes on methods, and wire up batch-processing and custom radii examples. Replace the previous CSV export section with a streamlined unnest+fwrite workflow for extracting coordination 
[…]
