+++ packages/crystract/vignettes/crystract.Rmd
@@ -47,7 +47,7 @@ Due to licensing restrictions, the example CIF files from the ICSD are no longer
 3.  **Store the File**: Save the file as `ICSD422.cif` in a known location on your computer.
 4.  **Update Paths**: In the code examples below, you will need to replace the placeholder path `"path/to/your/ICSD422.cif"` with the actual path to your downloaded file.
 
-**To run the code chunks yourself, you must also remove the `eval=FALSE` option from the chunk headers.**
+**To run the code chunks yourself, you must also remove the `eval=TRUE` option from the chunk headers.**
 
 ### Setup: Loading the Package
 
@@ -55,6 +55,7 @@ First, we load the `crystract` package.
 
 ```{r setup, message=FALSE, warning=FALSE}
 library(crystract)
+library(data.table)
 ```
 
 ## 1. The Core `crystract` Workflow
@@ -66,66 +67,44 @@ leverage the main wrapper function for an end-to-end pipeline.
 
 ### 1.1 The Full Pipeline for Batch Processing
 
-The `analyze_cif_files()` function is the cornerstone of the package. It
-performs the entire sequence of operations—from reading files to
-calculating bond angles with error propagation—for one or more files. It
-takes a vector of file paths and returns a single `data.table` where
-each row corresponds to a processed CIF file. Complex results (like
-coordinates, distances, and angles) are stored in list-columns for easy
-access and analysis.
+The `analyze_cif_files()` function (and its single-file counterpart `analyze_single_cif()`) is the cornerstone of the package. It performs the entire sequence of operations—from reading files to calculating bond angles with error propagation. You can easily specify which bonding algorithms to run simultaneously.
 
-This is the recommended function for batch processing.
+```{r full-pipeline-with-comments, eval=TRUE}
+# IMPORTANT: Update this path to point to your own downloaded CIF file.
+cif_path <- system.file("extdata", "1590946.cif", package = "crystract")
 
-```{r full-pipeline-with-comments, eval=FALSE}
-# IMPORTANT: Update this path to point to your downloaded CIF file.
-cif_path <- "path/to/your/ICSD422.cif"
-
-# Run the pipeline on our single example file
-analysis_results <- analyze_single_cif(cif_path)
-
-# You may also use analyze_cif_files() to process multiple files at once:
-# analysis_results <- analyze_cif_files(c("file1.cif", "file2.cif", ...))
-# Or a whole directory:
-# analysis_results <- analyze_cif_files("path/to/cif_directory/")
+# Run the pipeline on our single example file, extracting multiple bonding types at once
+analysis_results <- analyze_single_cif(
+  cif_path,
+  bonding_algorithms = c("minimum_distance", "brunner", "econ", "voronoi", "crystal_nn")
+)
 
 # Let's inspect the structure of the output table.
 # It's a single row containing all our results in nested data.tables.
 str(analysis_results, max.level = 2)
 ```
 
-As shown in the structure output, the result is a tidy `data.table`
-containing all extracted and calculated information. We can easily
-access the nested data frames for further analysis. Note that the output
-includes separate results for each bonding algorithm requested, such as
-`bonded_pairs_minimum_distance`, `bonded_pairs_brunner`, and `bonded_pairs_hoppe`.
[…]
