### 2026-02-27 480baa6
Remove clathrates vignette and docs link
Delete the clathrates vignette (packages/crystract/vignettes/clathrates.Rmd) and remove its entry from the pkgdown navbar. Update _pkgdown.yml to drop the "Example Use Case"/articles/clathrates.html link and tidy the reference section newline. This cleans up documentation by removing the deprecated/removed example vignette and its site navigation entry.
