+++ packages/crystract/vignettes/clathrates.Rmd
@@ -1,286 +1,0 @@
----
-title: "Advanced Workflow: High-Throughput Analysis of Clathrate Structures"
-author: "Don Ngo, Julia Maria Hubner, Anirudh Prabhu"
-date: "`r Sys.Date()`"
-output:
-  rmarkdown::html_vignette:
-vignette: >
-  %\VignetteIndexEntry{Advanced Workflow: High-Throughput Analysis of Clathrate Structures}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include=FALSE}
-# This chunk sets up the environment for the analysis.
-knitr::opts_chunk$set(
-  echo = TRUE, message = FALSE, warning = FALSE,
-  fig.width = 10, fig.height = 7, fig.align = 'center'
-)
-
-# Load all necessary libraries
-library(crystract)
-library(data.table)
-library(ggplot2)
-library(plotly)
-library(DT)
-library(htmlwidgets)
-library(knitr)
-
-# Load the package's built-in covalent radii data
-data(covalent_radii, package = "crystract")
-```
-
-## 1. Introduction and Objective
-
-This vignette demonstrates an advanced, robust workflow using the `crystract` package to analyze a large, complex dataset. While the introductory vignette focuses on the step-by-step analysis of a single, well-ordered crystal, this tutorial tackles the challenges of real-world, high-throughput analysis, specifically for disordered structures like clathrates.
-
-The goal is to process a dataset of clathrate structures to calculate a representative "weighted average network bond distance" for each one and explore its relationship with the lattice parameter 'a'.
-
-### The Robust Pipeline for Disordered Structures
-
-Clathrate structures often feature site-occupancy disorder and guest atoms within a host framework. A naive distance calculation can lead to non-physical "ghost" distances. To handle this, we will implement a specific, multi-step filtering pipeline for each crystal:
-
-1.  **Filter Guests:** Isolate the host framework by removing guest atoms.
-2.  **Filter Ghosts:** Remove physically implausible distances using `filter_ghost_distances()`.
-3.  **Identify Bonds:** Determine the true bonded pairs from the cleaned data using `minimum_distance()`.
-4.  **Calculate Average:** Compute the final `calculate_weighted_average_network_distance()` only on the confirmed, physical bonds.
-
-This **Guests -> Ghosts -> Bonds -> Average** sequence ensures the final metric is accurate and physically meaningful.
-
-## 2. Preparing the Data
-
-### 2.1. Obtaining and Organizing ICSD Files
-
-Due to licensing restrictions, the dataset of clathrate CIF files from the Inorganic Crystal Structure Database (ICSD) is **not** bundled with this package. To reproduce this workflow, you must download the data via your institutional license and organize it specifically for this script.
-
-#### Step 1: Download Required Files
-A list of the specific ICSD Collection Codes required for this analysis (and their corresponding structural categories) can be found in `required_ICSD_files.csv`, located in the root of the `ml-crystals` repository. Download the `.cif` file for each entry.
-
-#### Step 2: Create the Directory Structure
-The analysis script relies on the folder names to determine the **structural category** (which determines the Wyckoff sites used for calculations). You must create a main directory (e.g., `my_clathrate_data`) and create subdirectories named exactly after the categories found in the CSV file.
-
-The structure must look like this:
-
[…]
