+++ packages/crystract/_pkgdown.yml
@@ -11,8 +11,6 @@ navbar:
     menu:
     - text: Package Guide
       href: articles/crystract.html
-    - text: Example Use Case
-      href: articles/clathrates.html
   right:
   - icon: fa-github
     href: https://github.com/PrabhuLab/ml-crystals
