### 2026-02-27 3d9d0e3
Document calculate_weighted_neighbor_counts in pkgdown
Add calculate_weighted_neighbor_counts to the reference list in packages/crystract/_pkgdown.yml so the function is included in the generated pkgdown documentation. This is a documentation-only change to update the package reference index.
