### 2026-02-27 5c67201
Use pauling_en for electronegativity lookup
