### 2026-02-27 854b7f7
Register globals and fix PDF vignette output
