### 2026-02-27 3a762f8
Make vignette CIF path resolution robust
