### 2026-02-27 c17ec45
Rebuilt package vignettes
