+++ packages/crystract/doc/crystract.R
@@ -1,41 +1,47 @@
-## ----include = FALSE----------------------------------------------------------
+## ----include = FALSE-----------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-
-options(width = 80)
-
+options(width = 75)
 options(datatable.print.width = 80)
 
-## ----setup--------------------------------------------------------------------
+## ----setup, message=FALSE, warning=FALSE---------------------------------
 library(crystract)
+library(data.table)
 
-## ----full-pipeline-with-comment-----------------------------------------------
-# Find the path to the single example CIF file included in the package
-cif_path <- system.file("extdata", "ICSD422.cif", package = "crystract")
+## ----full-pipeline-with-comments, eval=TRUE------------------------------
+# IMPORTANT: Update this path to point to your own downloaded CIF file.
+# 1. Try to find the file in the installed package
+cif_path <- system.file("extdata", "1590946.cif", package = "crystract")
 
-# --- Example for batch processing a directory ---
-# To load all CIF files from the package's example directory, you would 
-# uncomment and run this:
-# extdata_dir <- system.file("extdata", package = "crystract")
-# all_cif_paths <- list.files(path = extdata_dir, pattern = "\\.cif$", full.names = TRUE)
-# analysis_results_batch <- analyze_cif_files(all_cif_paths)
+# 2. If that fails, look in the source directory
+if (cif_path == "") {
+  cif_path <- "../inst/extdata/1590946.cif"
+}
 
-# For this vignette, we will run the pipeline on just our single example file
-analysis_results <- analyze_cif_files(cif_path)
+# 3. Final check to provide a clear error if both fail
+if (!file.exists(cif_path)) {
+  stop("Could not find 1590946.cif in installed package or inst/extdata/")
+}
+
+# Run the pipeline on our single example file, extracting multiple bonding types at once
+analysis_results <- analyze_single_cif(
+  cif_path,
+  bonding_algorithms = c("minimum_distance", "brunner", "econ", "voronoi", "crystal_nn")
+)
 
 # Let's inspect the structure of the output table.
 # It's a single row containing all our results in nested data.tables.
 str(analysis_results, max.level = 2)
 
-## ----access-nested-data-------------------------------------------------------
-# The result is a list-column, so we access the element with [[1]]
-final_bonds <- analysis_results$bonded_pairs[[1]]
+## ----access-nested-data, eval=FALSE--------------------------------------
+# # The result is a list-column, so we access the element with [[1]]
+# final_bonds <- analysis_results$bonds_minimum_distance[[1]]
+# 
+# print(head(final_bonds))
 
-print(head(final_bonds))
-
-## ----load-data----------------------------------------------------------------
+## ----load-data, eval=TRUE------------------------------------------------
 # The path was defined in the previous section:
 cif_data_list <- read_cif_files(cif_path)
 
@@ -48,7 +54,7 @@ knitr::kable(
   caption = "First 6 lines of the raw CIF data."
 )
 
-## ----extract-metadata---------------------------------------------------------
+## ----extract-metadata, eval=TRUE-----------------------------------------
 database_code <- extract_database_code(cif_content)
[…]
