+++ packages/crystract/doc/crystract.Rmd
@@ -1,15 +1,14 @@
 ---
-title: "crystract"
-output: 
-  rmarkdown::html_vignette: default
-  rmarkdown::pdf_document: default
+title: "Using the crystract Package for Crystallographic Analysis"
+author: "Don Ngo, Julia Maria Hubner, Anirudh Prabhu"
+date: "`r Sys.Date()`"
+output:
+  rmarkdown::html_vignette:
+    toc: true
 vignette: >
-  %\VignetteIndexEntry{crystract}
+  %\VignetteIndexEntry{Using the crystract Package for Crystallographic Analysis}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
-editor_options: 
-  markdown: 
-    wrap: 72
 ---
 
 ```{r, include = FALSE}
@@ -17,9 +16,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-
-options(width = 80)
-
+options(width = 75)
 options(datatable.print.width = 80)
 ```
 
@@ -39,12 +36,24 @@ through the analysis of a single file step-by-step, explaining the
 underlying functions, the crystallographic principles, and the
 mathematical formulas they employ.
 
+### A Note on Example Data
+
+Due to licensing restrictions, the example CIF files from the ICSD are no longer bundled with the `crystract` package. To follow the examples in this vignette, you must download the required data yourself.
+
+1.  **Find Required Files**: A complete list of necessary files can be found in the `required_ICSD_files.csv` file at the root of the `ml-crystals` repository. For this introductory vignette, you will need **ICSD #422**.
+2.  **Download Data**: Access the [ICSD database](https://icsd.fiz-karlsruhe.de/) (requires a license) and download the `.cif` file for entry 422.
+3.  **Store the File**: Save the file as `ICSD422.cif` in a known location on your computer.
+4.  **Update Paths**: In the code examples below, you will need to replace the placeholder path `"path/to/your/ICSD422.cif"` with the actual path to your downloaded file.
+
+**To run the code chunks yourself, you must also remove the `eval=TRUE` option from the chunk headers.**
+
 ### Setup: Loading the Package
 
 First, we load the `crystract` package.
 
-```{r setup}
+```{r setup, message=FALSE, warning=FALSE}
 library(crystract)
+library(data.table)
 ```
 
 ## 1. The Core `crystract` Workflow
@@ -56,72 +65,55 @@ leverage the main wrapper function for an end-to-end pipeline.
 
 ### 1.1 The Full Pipeline for Batch Processing
 
-The `analyze_cif_files()` function is the cornerstone of the package. It
-performs the entire sequence of operations—from reading files to
-calculating bond angles with error propagation—for one or more files. It
-takes a vector of file paths and returns a single `data.table` where
-each row corresponds to a processed CIF file. Complex results (like
-coordinates, distances, and angles) are stored in list-columns for easy
-access and analysis.
+The `analyze_cif_files()` function (and its single-file counterpart `analyze_single_cif()`) is the cornerstone of the package. It performs the entire sequence of operations—from reading files to calculating bond angles with error propagation. You can easily specify which bonding algorithms to run simultaneously.
 
-This is the recommended function for batch processing. The code block
-below shows a commented-out example of how you would load all CIF files
-from the package's example data directory. For this demonstration, we
-will proceed by analyzing just a single example file.
[…]
