+++ articles/index.md
@@ -2,7 +2,5 @@
 
 ### All vignettes
 
-- [Advanced Workflow: High-Throughput Analysis of Clathrate
-  Structures](https://prabhulab.github.io/ml-crystals/articles/clathrates.md):
 - [Using the crystract Package for Crystallographic
   Analysis](https://prabhulab.github.io/ml-crystals/articles/crystract.md):
