+++ llms.txt
@@ -352,6 +352,8 @@ Functions for geometry and crystallography calculations.
   : Calculate Bond Angles
 - [`calculate_neighbor_counts()`](https://prabhulab.github.io/ml-crystals/reference/calculate_neighbor_counts.md)
   : Calculate Coordination Numbers
+- [`calculate_weighted_neighbor_counts()`](https://prabhulab.github.io/ml-crystals/reference/calculate_weighted_neighbor_counts.md)
+  : Calculate Weighted Coordination Numbers
 - [`calculate_weighted_average_network_distance()`](https://prabhulab.github.io/ml-crystals/reference/calculate_weighted_average_network_distance.md)
   : Calculate Weighted Average Network Bond Distance
 
@@ -441,7 +443,5 @@ Parsers for specific CIF data blocks.
 
 ### All vignettes
 
-- [Advanced Workflow: High-Throughput Analysis of Clathrate
-  Structures](https://prabhulab.github.io/ml-crystals/articles/clathrates.md):
 - [Using the crystract Package for Crystallographic
   Analysis](https://prabhulab.github.io/ml-crystals/articles/crystract.md):
