### 2026-02-27 9928090
Remove ICSD example-data instructions from vignette
