+++ packages/crystract/vignettes/crystract.Rmd
@@ -22,30 +22,9 @@ options(datatable.print.width = 80)
 
 ## Introduction
 
-This vignette demonstrates the core workflow for using the `crystract`
-package. The package is designed for the **batch processing** of
-Crystallographic Information Files (.cif) to extract structural and
-chemical information from the repeating atomic arrangement that defines
-a crystal. It calculates key geometric properties like bond lengths and
-angles and compiles the results into a structured format suitable for
-large-scale data analysis.
+This vignette demonstrates the core workflow for using the `crystract` package. The package is designed for the **batch processing** of Crystallographic Information Files (.cif) to extract structural and chemical information from the repeating atomic arrangement that defines a crystal. It calculates key geometric properties like bond lengths and angles and compiles the results into a structured format suitable for large-scale data analysis.
 
-We will first demonstrate the main batch-processing function,
-`analyze_cif_files`. Then, to illustrate how it works, we will walk
-through the analysis of a single file step-by-step, explaining the
-underlying functions, the crystallographic principles, and the
-mathematical formulas they employ.
-
-### A Note on Example Data
-
-Due to licensing restrictions, the example CIF files from the ICSD are no longer bundled with the `crystract` package. To follow the examples in this vignette, you must download the required data yourself.
-
-1.  **Find Required Files**: A complete list of necessary files can be found in the `required_ICSD_files.csv` file at the root of the `ml-crystals` repository. For this introductory vignette, you will need **ICSD #422**.
-2.  **Download Data**: Access the [ICSD database](https://icsd.fiz-karlsruhe.de/) (requires a license) and download the `.cif` file for entry 422.
-3.  **Store the File**: Save the file as `ICSD422.cif` in a known location on your computer.
-4.  **Update Paths**: In the code examples below, you will need to replace the placeholder path `"path/to/your/ICSD422.cif"` with the actual path to your downloaded file.
-
-**To run the code chunks yourself, you must also remove the `eval=TRUE` option from the chunk headers.**
+We will first demonstrate the main batch-processing function, `analyze_cif_files`. Then, to illustrate how it works, we will walk through the analysis of a single file step-by-step, explaining the underlying functions, the crystallographic principles, and the mathematical formulas they employ.
 
 ### Setup: Loading the Package
 
@@ -58,10 +37,7 @@ library(data.table)
 
 ## 1. The Core `crystract` Workflow
 
-The primary goal of `crystract` is to automate the analysis of many CIF
-files at once. While the package provides granular functions for each
-step of the crystallographic analysis, the most common use case is to
-leverage the main wrapper function for an end-to-end pipeline.
+The primary goal of `crystract` is to automate the analysis of many CIF files at once. While the package provides granular functions for each step of the crystallographic analysis, the most common use case is to leverage the main wrapper function for an end-to-end pipeline.
 
 ### 1.1 The Full Pipeline for Batch Processing
 
@@ -110,8 +86,7 @@ To understand what happens under the hood, this section breaks down the process 
 
 #### 1.2.1 Loading CIF Data
 
-We use the package's `read_cif_files()` function to load data into
[…]
