+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   crystract: crystract.html
-last_built: 2026-02-27T03:57Z
+last_built: 2026-02-27T04:14Z
 urls:
   reference: https://prabhulab.github.io/ml-crystals/reference
   article: https://prabhulab.github.io/ml-crystals/articles
