### 2026-02-27 ee060f6
Merge partial-occupancy sites for tessellation
