+++ reference/extract_space_group_name.md
@@ -29,3 +29,14 @@ Other extractors:
 [`extract_structure_type()`](https://prabhulab.github.io/ml-crystals/reference/extract_structure_type.md),
 [`extract_symmetry_operations()`](https://prabhulab.github.io/ml-crystals/reference/extract_symmetry_operations.md),
 [`extract_unit_cell_metrics()`](https://prabhulab.github.io/ml-crystals/reference/extract_unit_cell_metrics.md)
+
+## Examples
+
+``` r
+cif_path <- system.file("extdata", "1590946.cif", package = "crystract")
+if (file.exists(cif_path)) {
+  cif_content <- read_cif_files(cif_path)[[1]]
+  extract_space_group_name(cif_content)
+}
+#> [1] "P n m a"
+```
