+++ authors.md
@@ -2,18 +2,18 @@
 
 ## Authors
 
-- **Don Ngo**. Maintainer.
+- **Don Ngo**. Author, maintainer.
 
 ## Citation
 
 Ngo D (2026). *crystract: Crystallographic Information File (CIF) Data
-Processing Tools*. R package version 1.0.1,
+Processing Tools*. R package version 1.0.0,
 <https://prabhulab.github.io/ml-crystals/>.
 
     @Manual{,
       title = {crystract: Crystallographic Information File (CIF) Data Processing Tools},
       author = {Don Ngo},
       year = {2026},
-      note = {R package version 1.0.1},
+      note = {R package version 1.0.0},
       url = {https://prabhulab.github.io/ml-crystals/},
     }
