### 2026-03-03 3f718ea
Enhance documentation: refs, diagrams, README
Add bibliographic references to the package and bonding functions, and expand documentation with workflow/decision diagrams and updated examples. Updates include: adding @references blocks in R/crystract-package.R and R/predict_bonds.R for multiple bonding algorithms; major README.Rmd and generated README.html revisions (repo path/badge fixes, added Voronoi/CrystalNN mentions, new quickstart/demo using analyze_single_cif, expanded data dictionary and outputs); updated man pages; added man/figures (bonding_logic.png, radii_logic.png, workflow_diagra
[…]
