+++ packages/crystract/R/crystract-package.R
@@ -9,6 +9,16 @@
 #' angles. Facilitates the preparation of crystallographic data for further
 #' analysis, including machine learning applications in materials science.
 #'
+#' @references
+#' Shannon, R. D. (1976). "Revised effective ionic radii and systematic studies of
+#' interatomic distances in halides and chalcogenides." \emph{Acta Crystallographica
+#' Section A: Crystal Physics, Diffraction, Theoretical and General Crystallography},
+#' 32(5), 751-767.
+#'
+#' Pauling, L. (1960). \emph{The Nature of the Chemical Bond and the Structure of
+#' Molecules and Crystals: An Introduction to Modern Structural Chemistry} (Vol. 18).
+#' Cornell University Press.
+#'
 #' @keywords internal
 "_PACKAGE"
 #'
