+++ packages/crystract/README.md
@@ -9,7 +9,7 @@
 
 <!-- [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental) -->
 
-<!-- [![R-CMD-check](https://github.com/Don-Tat/ml-crystals/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/Don-Tat/ml-crystals/actions/workflows/R-CMD-check.yaml) -->
+<!-- [![R-CMD-check](https://github.com/PrabhuLab/ml-crystals/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/PrabhuLab/ml-crystals/actions/workflows/R-CMD-check.yaml) -->
 
 <!-- badges: end -->
 
@@ -18,17 +18,17 @@ Information Files (`.cif`), extracting essential data such as chemical
 formulas, unit cell parameters, atomic coordinates, and symmetry
 operations. It also includes tools to calculate interatomic distances,
 identify bonded pairs using various algorithms (Minimum Distance,
-Brunner’s, Hoppe’s), determine nearest neighbor counts, and calculate
-bond angles. All data is extracted into nested data.tables, which can
-then be exported as an R Data Structure (RDS) or folders of .csv files.
-The package is designed to facilitate the preparation of
-crystallographic data for further analysis, including machine learning
-applications in materials science.
+Brunner’s, Hoppe’s, Voronoi, CrystalNN), determine nearest neighbor
+counts, and calculate bond angles. All data is extracted into nested
+`data.table`s, which can then be exported as an R Data Structure (RDS)
+or folders of .csv files. The package is designed to facilitate the
+preparation of crystallographic data for further analysis, including
+machine learning applications in materials science.
 
 > **Note on Repository Structure**
 >
-> The `crystract` package is located within the `crystract/`
-> subdirectory of the `Prabhulab/ml-crystals` GitHub repository. You
+> The `crystract` package is located within the `packages/crystract/`
+> subdirectory of the `PrabhuLab/ml-crystals` GitHub repository. You
 > must use the `subdir` argument during installation, as shown below.
 
 ## Key Features
@@ -36,174 +36,205 @@ applications in materials science.
 - **Efficient CIF Parsing**: Utilizes `data.table` for fast and robust
   extraction of metadata, unit cell parameters, atomic coordinates, and
   symmetry operations.
-
 - **Symmetry and Supercell Generation**: Applies symmetry operations to
   generate a full unit cell from the asymmetric unit and expands
   coordinates into a 3x3x3 supercell for neighbor searching.
-
 - **Geometric Calculations**: Computes interatomic distances using the
   metric tensor (correct for all crystal systems) and calculates bond
   angles.
-
 - **Multiple Bonding Algorithms**: Implements several algorithms to
   identify bonded atoms, including the `minimum_distance` (default),
-  `brunner`, and `hoppe` methods.
-
+  `brunner`, `econ` (Hoppe’s), `voronoi`, and `crystal_nn` methods.
 - **Rigorous Error Propagation**: Calculates and propagates experimental
   uncertainties from the CIF file into the final calculated bond lengths
   and angles.
-
 - **Powerful Post-Processing Tools**: Includes functions to filter
   results by chemical element, Wyckoff site, or to remove non-physical
   “ghost” distances caused by site disorder using a customizable atomic
   radii table.
-
 - **Batch Processing & Export**: The main `analyze_cif_files()` function
[…]
