+++ packages/crystract/man/crystal_nn.Rd
@@ -50,6 +50,19 @@ if (file.exists(cif_path)) {
 }
 }
 }
+\references{
+Shannon, R. D. (1976). "Revised effective ionic radii and systematic studies of
+interatomic distances in halides and chalcogenides." \emph{Acta Crystallographica
+Section A}, 32(5), 751-767.
+
+Pauling, L. (1960). \emph{The Nature of the Chemical Bond and the Structure of
+Molecules and Crystals: An Introduction to Modern Structural Chemistry}.
+Cornell University Press.
+
+Pan, H., Ganose, A. M., Horton, M., et al. (2021). "Benchmarking Coordination Number
+Prediction Algorithms on Inorganic Crystal Structures." \emph{Inorganic Chemistry},
+60(3), 1590–1603. \doi{10.1021/acs.inorgchem.0c02996}
+}
 \seealso{
 Other bonding algorithms: 
 \code{\link{brunner_nn_reciprocal}()},
