+++ packages/crystract/man/voronoi_nn.Rd
@@ -38,6 +38,13 @@ if (file.exists(cif_path)) {
 }
 }
 }
+\references{
+O'Keeffe, M. (1979). "A Proposed Rigorous Definition of Coordination Number."
+\emph{Acta Crystallographica Section A}, 35(5), 772–775. \doi{10.1107/S0567739479001711}
+
+Aurenhammer, F., Klein, R., & Lee, D.-T. (2013). \emph{Voronoi Diagrams and
+Delaunay Triangulations}. World Scientific Publishing Company.
+}
 \seealso{
 Other bonding algorithms: 
 \code{\link{brunner_nn_reciprocal}()},

[…]
