### 2026-03-03 f0e1d23
Update install guide to use remotes
