+++ llms.txt
@@ -5,17 +5,17 @@ Information Files (`.cif`), extracting essential data such as chemical
 formulas, unit cell parameters, atomic coordinates, and symmetry
 operations. It also includes tools to calculate interatomic distances,
 identify bonded pairs using various algorithms (Minimum Distance,
-Brunner’s, Hoppe’s), determine nearest neighbor counts, and calculate
-bond angles. All data is extracted into nested data.tables, which can
-then be exported as an R Data Structure (RDS) or folders of .csv files.
-The package is designed to facilitate the preparation of
-crystallographic data for further analysis, including machine learning
-applications in materials science.
+Brunner’s, Hoppe’s, Voronoi, CrystalNN), determine nearest neighbor
+counts, and calculate bond angles. All data is extracted into nested
+`data.table`s, which can then be exported as an R Data Structure (RDS)
+or folders of .csv files. The package is designed to facilitate the
+preparation of crystallographic data for further analysis, including
+machine learning applications in materials science.
 
 > **Note on Repository Structure**
 >
-> The `crystract` package is located within the `crystract/`
-> subdirectory of the `Prabhulab/ml-crystals` GitHub repository. You
+> The `crystract` package is located within the `packages/crystract/`
+> subdirectory of the `PrabhuLab/ml-crystals` GitHub repository. You
 > must use the `subdir` argument during installation, as shown below.
 
 ## Key Features
@@ -23,28 +23,22 @@ applications in materials science.
 - **Efficient CIF Parsing**: Utilizes `data.table` for fast and robust
   extraction of metadata, unit cell parameters, atomic coordinates, and
   symmetry operations.
-
 - **Symmetry and Supercell Generation**: Applies symmetry operations to
   generate a full unit cell from the asymmetric unit and expands
   coordinates into a 3x3x3 supercell for neighbor searching.
-
 - **Geometric Calculations**: Computes interatomic distances using the
   metric tensor (correct for all crystal systems) and calculates bond
   angles.
-
 - **Multiple Bonding Algorithms**: Implements several algorithms to
   identify bonded atoms, including the `minimum_distance` (default),
-  `brunner`, and `hoppe` methods.
-
+  `brunner`, `econ` (Hoppe’s), `voronoi`, and `crystal_nn` methods.
 - **Rigorous Error Propagation**: Calculates and propagates experimental
   uncertainties from the CIF file into the final calculated bond lengths
   and angles.
-
 - **Powerful Post-Processing Tools**: Includes functions to filter
   results by chemical element, Wyckoff site, or to remove non-physical
   “ghost” distances caused by site disorder using a customizable atomic
   radii table.
-
 - **Batch Processing & Export**: The main
   [`analyze_cif_files()`](https://prabhulab.github.io/ml-crystals/reference/analyze_cif_files.md)
   function is designed to process hundreds of files in a single run, and
@@ -52,152 +46,178 @@ applications in materials science.
   with
   [`export_analysis_to_csv()`](https://prabhulab.github.io/ml-crystals/reference/export_analysis_to_csv.md).
 
-### Here is a comprehensive Data Dictionary for the **crystract** package.
+------------------------------------------------------------------------
 
-It is organized into three sections: 1. **Master Analysis Object**: The
-main output structure returned by
-[`analyze_cif_files()`](https://prabhulab.github.io/ml-crystals/reference/analyze_cif_files.md)
-and
-[`analyze_single_cif()`](https://prabhulab.github.io/ml-crystals/reference/analyze_single_cif.md).
-2. **Nested Data Tables**: The specific structure of the sub-tables
-contained within the list-columns of the Master Object. 3. **Reference
-Data**: The structure of auxiliary tables like atomic radii.
+## Workflow Overview
+
+The following diagram illustrates the primary data pipeline in
[…]
