+++ reference/export_analysis_to_csv.md
@@ -39,6 +39,6 @@ if (file.exists(cif_path)) {
 
   unlink(out_dir, recursive = TRUE)
 }
-#> Analysis successfully exported to: /tmp/RtmpmXXsA2/cif_csvs
+#> Analysis successfully exported to: /tmp/RtmplU5NFq/cif_csvs
 # }
 ```
