+++ reference/crystract-package.md
@@ -9,6 +9,17 @@ counts, and calculate bond angles. Facilitates the preparation of
 crystallographic data for further analysis, including machine learning
 applications in materials science.
 
+## References
+
+Shannon, R. D. (1976). "Revised effective ionic radii and systematic
+studies of interatomic distances in halides and chalcogenides." *Acta
+Crystallographica Section A: Crystal Physics, Diffraction, Theoretical
+and General Crystallography*, 32(5), 751-767.
+
+Pauling, L. (1960). *The Nature of the Chemical Bond and the Structure
+of Molecules and Crystals: An Introduction to Modern Structural
+Chemistry* (Vol. 18). Cornell University Press.
+
 ## See also
 
 Useful links:
