+++ reference/econ_nn.md
@@ -31,6 +31,18 @@ econ_nn(distances, atomic_coordinates, tol = 0.2, use_fictive_radius = FALSE)
 
 A `data.table` of bonded pairs.
 
+## References
+
+Hoppe, R. (1979). "Effective Coordination Numbers (ECoN) and Mean
+Fictive Ionic Radii (MEFIR)." *Zeitschrift Für Kristallographie*,
+150(1–4), 23–52.
+[doi:10.1524/zkri.1979.150.14.23](https://doi.org/10.1524/zkri.1979.150.14.23)
+
+Shannon, R. D. (1976). "Revised Effective Ionic Radii and Systematic
+Studies of Interatomic Distances in Halides and Chalcogenides." *Acta
+Crystallographica Section A*, 32(5), 751–767.
+[doi:10.1107/S0567739476001551](https://doi.org/10.1107/S0567739476001551)
+
 ## See also
 
 Other bonding algorithms:
