+++ authors.md
@@ -8,12 +8,12 @@
 
 Ngo D (2026). *crystract: Crystallographic Information File (CIF) Data
 Processing Tools*. R package version 1.0.0,
-<https://prabhulab.github.io/ml-crystals/>.
+<https://prabhulab.github.io/ml-crystals/packages/crystract>.
 
     @Manual{,
       title = {crystract: Crystallographic Information File (CIF) Data Processing Tools},
       author = {Don Ngo},
       year = {2026},
       note = {R package version 1.0.0},
-      url = {https://prabhulab.github.io/ml-crystals/},
+      url = {https://prabhulab.github.io/ml-crystals/packages/crystract},
     }
