### 2026-03-03 #9529 @Don-Tat
Hi @zhubonan

Other than the fact that we have submitted to CRAN, no other updates. Perhaps while waiting for CRAN approval release, we can move on forward with the review. With this CRAN submission I believe we have addressed all the existing comments from @bobleesj and can restart the review process. We will let you know once we receive approval, but this should clear the requirements for package installation without any compilers required.

Thanks for your patience @zhubonan, @bobleesj and @singularitti.
