### 2026-03-10 75808fd
Update DESCRIPTION: authors, license, algorithms
