+++ authors.md
@@ -3,17 +3,23 @@
 ## Authors
 
 - **Don Ngo**. Author, maintainer.
+  [](https://orcid.org/0009-0001-2779-2146)
+
+- **Anirudh Prabhu**. Author. [](https://orcid.org/0000-0002-9921-6084)
+
+- **Julia-Maria Hubner**. Author.
+  [](https://orcid.org/0000-0003-2048-6629)
 
 ## Citation
 
-Ngo D (2026). *crystract: Crystallographic Information File (CIF) Data
-Processing Tools*. R package version 1.0.0,
-<https://prabhulab.github.io/ml-crystals/packages/crystract>.
+Ngo D, Prabhu A, Hubner J (2026). *crystract: Crystallographic
+Information File (CIF) Data Processing Tools*. R package version 1.0.0,
+<https://prabhulab.github.io/ml-crystals/>.
 
     @Manual{,
       title = {crystract: Crystallographic Information File (CIF) Data Processing Tools},
-      author = {Don Ngo},
+      author = {Don Ngo and Anirudh Prabhu and Julia-Maria Hubner},
       year = {2026},
       note = {R package version 1.0.0},
-      url = {https://prabhulab.github.io/ml-crystals/packages/crystract},
+      url = {https://prabhulab.github.io/ml-crystals/},
     }
