### 2026-03-17 d37f5d8
Improve vignette options, prompt, and docs
