+++ notebook/natural/naturalTest.Rmd
@@ -21,11 +21,13 @@ knitr::opts_chunk$set(echo = TRUE, message = TRUE, warning = FALSE)
 
 ## 1. Introduction
 
-This workflow performs a targeted analysis of Sulfides, Selenides, and Tellurides. It utilizes the Minimum Distance bonding algorithm within the parallelized extraction pipeline and includes diagnostic steps for high Coordination Numbers (CN > 12) and Weighted Coordination Numbers.
+This workflow performs a targeted, large-scale crystallographic analysis of Sulfides, Selenides, and Tellurides. By utilizing the Minimum Distance bonding algorithm within the highly parallelized `crystract` extraction pipeline, we calculate fundamental geometric metrics such as bond distances, bond angles, and coordination numbers (CN) across thousands of parsed `.cif` files.
 
-**Note:** All plots in this report are interactive. You can hover to see details, zoom in/out, and click legend items to filter data.
+Furthermore, this notebook introduces critical diagnostic steps to evaluate high-coordination environments (Geometric CN > 12) and to compute **Weighted Coordination Numbers**. Weighted CN provides a more accurate representation of the local chemical environment by taking into account crystallographic site occupancies and resolving partial occupancies inherent to structural disorder.
 
 ## 2. Setup: Libraries and Reference Data
+
+We begin by loading the necessary data manipulation, crystallography, and parallel processing packages. A robust periodic table reference (`ptable`) is established to ensure all visualizations can accurately sort elements by their atomic number (Z) rather than alphabetically, exposing true periodic trends.
 
 ```{r load-libraries, message=FALSE, warning=FALSE}
 # Data Manipulation
@@ -48,6 +50,7 @@ library(viridis)
 library(gridExtra)
 library(DT)      # For interactive tables
 library(plotly)  # For interactive plots
+library(scales)  # For plot scales and breaks
 
 # --- Initialize Time Tracking ---
 timing_log <- data.frame(
@@ -81,12 +84,11 @@ ptable <- data.table(
 
 ## 3. Data Discovery and Chemistry Filtering
 
-Filter criteria: **(Contains S OR Se OR Te) AND (Does NOT contain O)**.
+To analyze Chalcogenides strictly, our filtering logic is: **(Contains S OR Se OR Te) AND (Does NOT contain O)**. Excluding oxygen prevents the inclusion of sulfates, selenites, and tellurates, isolating the data strictly to sulfides, selenides, and tellurides.
 
 ```{r filter-by-chemistry}
 # --- Configuration ---
-# UPDATE THIS PATH to your local CIF directory
-cif_directory <- "/Users/carnegieuser/Documents/amcsd" 
+cif_directory <- "/Users/carnegieuser/Downloads/cif" 
 
 step_start <- Sys.time()
 
@@ -132,7 +134,7 @@ cat(sprintf("Filtering complete. Kept: %d files.\n", length(target_files)))
 
 ## 4. Parallel Analysis (Optimized)
 
-We now run `crystract` with `bonding_algorithms = "minimum_distance"` and `calculate_bond_angles = TRUE` using a customized robust worker function.
[…]
