### 2026-03-22 #9529 @Don-Tat
Hi @zhubonan, @bobleesj, @singularitti, and @hisacroc,

Our package has now been accepted and published by CRAN https://cran.rstudio.com/web/packages/crystract/index.html. This should address the installation requirements discussed earlier.

We can now restart/continue the review process. I look forward to working with the reviewers on any comments or suggestions they have.

Thank you for your time and consideration.

Best,
Don-Tat
