### 2026-03-23 #9529 @bobleesj
@Don-Tat Great. It works well with the package downloaded from CRAN. I tested using a cif from Materials Project CIF, aside from ICSD cif:
```
  > result$bonds_minimum_distance[[1]]
       Atom1        Atom2 Distance     DeltaX     DeltaY     DeltaZ    Weight DistanceError
   1:    Co0 O11_1_0_-1_0 1.912229  0.2108378  0.2630569 -0.2369476 0.9999820             0
   2:    Co0  O12_1_0_0_0 1.912229  0.2369523 -0.2630502  0.2369449 0.9999819             0
   3:    Co0 O13_1_0_-1_0 1.912194 -0.2369414  0.2630540  0.2108330 1.0000000             0
   4:    Co0  O6_1_0_-1_0 1.912205 -0.2369432  0.2630539 -0.2369474 0.9999941             0
   5:    Co0   O7_1_0_0_0 1.912216  0.2369505 -0.2630503 -0.2108430 0.9999885             0
   6:    Co0   O9_1_0_0_0 1.912208 -0.2108493 -0.2630418  0.2369459 0.9999926             0
   7:    Co1 O10_1_-1_0_0 1.912211  0.2630545  0.2108364 -0.2369464 0.9999955             0
   8:    Co1  O12_1_0_0_0 1.912220 -0.2630477  0.2369498  0.2369449 0.9999905             0
   9:    Co1 O13_1_-1_0_0 1.912209  0.2630586 -0.2369460  0.2108330 0.9999964             0
  10:    Co1  O6_1_-1_0_0 1.912220  0.2630568 -0.2369461 -0.2369474 0.9999909             0
  11:    Co1   O7_1_0_0_0 1.912208 -0.2630495  0.2369497 -0.2108430 0.9999968             0
  12:    Co1   O8_1_0_0_0 1.912202 -0.2630472 -0.2108421  0.2369458 1.0000000             0
  13:    Co2 O10_1_0_0_-1 1.912208 -0.2369455  0.2108364  0.2630536 0.9999989             0
  14:    Co2 O11_1_0_0_-1 1.912206  0.2108378 -0.2369431  0.2630524 1.0000000             0
  15:    Co2  O12_1_0_0_0 1.912256  0.2369523  0.2369498 -0.2630551 0.9999740             0
  16:    Co2  O6_1_0_0_-1 1.912207  0.2369432  0.2369461 -0.2630526 0.9999999             0
  17:    Co2   O8_1_0_0_0 1.912233  0.2369528 -0.2108421 -0.2630542 0.9999860             0
  18:    Co2   O9_1_0_0_0 1.912257 -0.2108493  0.2369582 -0.2630541 0.9999733
```
and angles info
```
  > result$angles_minimum_distance[[1]]
      CentralAtom    Neighbor1    Neighbor2     Angle AngleError
   1:         Co0 O11_1_0_-1_0  O12_1_0_0_0  96.46109          0
   2:         Co0 O11_1_0_-1_0 O13_1_0_-1_0  83.53765          0
   3:         Co0 O11_1_0_-1_0  O6_1_0_-1_0  83.53732          0
   4:         Co0 O11_1_0_-1_0   O7_1_0_0_0  96.46075          0
   5:         Co0 O11_1_0_-1_0   O9_1_0_0_0 179.99777          0
   6:         Co0  O12_1_0_0_0 O13_1_0_-1_0  96.46043          0
   7:         Co0  O12_1_0_0_0  O6_1_0_-1_0 179.99841          0
   8:         Co0  O12_1_0_0_0   O7_1_0_0_0  83.53979          0
   9:         Co0  O12_1_0_0_0   O9_1_0_0_0  83.54056          0
```

A few checklist comments:

- Tests are written.

```
Input:
  devtools::test()  # from cloned source repo
Output:
  ✔ |  19 | 01-extractors
  ✔ |   5 | 02-coord-processing
  ✔ |  16 | 03-property-calculations
  ✔ |  10 | 04-post-processing
  ✔ |  12 | 05-main-pipeline
```

- Algorithms such as `crystal_nn` and `voronoi` are implemented as described.
- `CONTRIBUTING.md` is provided.

A few more comments for authors if you will:

- [x] DOI link needs a fix:

- [x] References for ICSD or COD, Voronoi, and including CrystalNN, MinimumDistanceNN, etc. Please double-check any missing references. Hope @anirudhprabhu can help with this.
- [x] CRAN installation in [README](https://github.com/PrabhuLab/ml-crystals/blob/main/packages/crystract/README.md)?

- [x] The "Experimental uncertainties" claims are used in the paper about 5 times and it is one of the two differentiating factors (along with bond angle calculations) but I haven't yet really seen what "experimental uncertainties" are referring to both in the paper and the documentation.

Overall, great improvement, with documentation, no need to compile run the code, example outputs, etc.
