### 2026-03-25 6584786
Fix DOI format in paper.bib
Updated DOI format to remove the URL scheme.
