+++ paper/paper.bib
@@ -17,7 +17,7 @@
   pages={314--319},
   year={2013},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.commatsci.2012.10.028}
+  doi = {10.1016/j.commatsci.2012.10.028}
 }
 
 @article{Pan2021,
