+++ news/index.md
@@ -2,4 +2,6 @@
 
 ## crystract 1.0.0
 
+CRAN release: 2026-03-20
+
 - Initial CRAN submission.
