### 2026-03-25 698bdc8
Update crystract installation instructions
