### 2026-03-30 a98dd75
Add multiple articles and books to bibliography
