+++ paper/paper.bib
@@ -150,3 +150,71 @@
   year = {2024},
   doi = {10.1016/j.jallcom.2023.173241}
 }
+
+@article{Hoppe1979,
+	title = {Effective coordination numbers ({ECoN}) and mean fictive ionic radii ({MEFIR})},
+	volume = {150},
+	url = {https://doi.org/10.1524/zkri.1979.150.14.23},
+	doi = {doi:10.1524/zkri.1979.150.14.23},
+	number = {1-4},
+	urldate = {2026-02-24},
+	journal = {Zeitschrift für Kristallographie - Crystalline Materials},
+	author = {Hoppe, Rudolf},
+	year = {1979},
+	pages = {23--52},
+}
+
+@article{Brunner1977,
+	title = {A definition of coordination and its relevance in the structure types {AlB}{\textbackslash}sb 2 and {NiAs}},
+	volume = {33},
+	url = {https://doi.org/10.1107/S0567739477000461},
+	doi = {10.1107/S0567739477000461},
+	abstract = {The largest gap in the list of reciprocal interatomic distances is used to limit the coordination polyhedra. For the structure types AIB{\textbackslash}sb 2 and NiAs, the polyhedra are evaluated according to ıt c/ıt a. In plots of ıt c/ıt a ıt versus ıt r{\textbackslash}sb A/ıt r{\textbackslash}sb B, the various coordination types correspond to actually observed families of representatives. These results as well as others from more than 50 structure types support the view that coordination as limited by the largest gap is a realistic factor.},
+	number = {1},
+	journal = {Acta Crystallographica Section A},
+	author = {Brunner, G. O.},
+	month = jan,
+	year = {1977},
+	pages = {226--227},
+}
+
+@article{Shannon1976,
+	title = {Revised effective ionic radii and systematic studies of interatomic distances in halides and chalcogenides},
+	volume = {32},
+	number = {5},
+	journal = {Foundations of Crystallography},
+	publisher = {International Union of Crystallography},
+	author = {Shannon, Robert D},
+	year = {1976},
+	pages = {751--767},
+}
+
+@book{Pauling1960,
+	title = {The nature of the chemical bond and the structure of molecules and crystals: an introduction to modern structural chemistry},
+	volume = {18},
+	publisher = {Cornell university press},
+	author = {Pauling, Linus},
+	year = {1960},
+}
+
+@book{AurenhammerVoronoi2013,
+	title = {Voronoi diagrams and {Delaunay} triangulations},
+	publisher = {World Scientific Publishing Company},
+	author = {Aurenhammer, Franz and Klein, Rolf and Lee, Der-Tsai},
+	year = {2013},
+}
+
+@article{Okeeffe1979,
+	title = {A proposed rigorous definition of coordination number},
+	volume = {35},
+	url = {https://doi.org/10.1107/S0567739479001765},
+	doi = {10.1107/S0567739479001765},
+	abstract = {It is proposed that a weighted coordination number, ıt Z*, be used as a measure of the number of neighbours of an atom in a crystal. Coordinating atoms contribute faces to the Voronoi polyhedron around a central atom and their contributions are weighted in proportion to the solid angle subtended by that face at the centre. The advantages of this definition over other proposals are pointed out.},
+	number = {5},
+	journal = {Acta Crystallographica Section A},
+	author = {O'Keeffe, M.},
+	month = sep,
+	year = {1979},
+	pages = {772--775},
+}
+
