### 2026-04-07 675098f
Refactor CIF extraction, update example CIF & paper
