+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={6},
   pages={655--685},
   year={1991},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
+  doi={10.1107/S010876739101067X}
 }
 
 @article{Ong2013,
@@ -17,7 +18,7 @@
   pages={314--319},
   year={2013},
   publisher={Elsevier},
-  doi = {10.1016/j.commatsci.2012.10.028}
+  doi={10.1016/j.commatsci.2012.10.028}
 }
 
 @article{Pan2021,
@@ -29,7 +30,7 @@
   pages={1590--1603},
   year={2021},
   publisher={ACS Publications},
-  doi = {10.1021/acs.inorgchem.0c02996}
+  doi={10.1021/acs.inorgchem.0c02996}
 }
 
 @article{Groom2016,
@@ -41,18 +42,18 @@
   pages={171--179},
   year={2016},
   publisher={International Union of Crystallography},
-  doi = {10.1107/s2052520616003954}
+  doi={10.1107/s2052520616003954}
 }
 
 @article{Macrae2006,
-  title={Mercury: visualization and analysis of crystal structures},
-  author={Macrae, Clare F and Edgington, Paul R and McCabe, Patrick and Pidcock, Elna and Shields, Greg P and Taylor, Robin and Towler, Matthew and Streek, JVD},
-  journal={Applied Crystallography},
-  volume={39},
-  number={3},
-  pages={453--457},
-  year={2006},
-  publisher={International Union of Crystallography},
+  author = "Macrae, Clare F. and Edgington, Paul R. and McCabe, Patrick and Pidcock, Elna and Shields, Greg P. and Taylor, Robin and Towler, Matthew and van de Streek, Jacco",
+  title = "{{\it Mercury}: visualization and analysis of crystal structures}",
+  journal = "Journal of Applied Crystallography",
+  year = "2006",
+  volume = "39",
+  number = "3",
+  pages = "453--457",
+  month = "Jun",
   doi = {10.1107/S002188980600731X}
 }
 
@@ -64,7 +65,7 @@
   number={2},
   pages={335--338},
   year={2004},
-  doi = {10.1107/S0021889804003528}
+  doi={10.1107/S0021889804003528}
 }
 
 @Manual{Foadi2025,
@@ -72,7 +73,7 @@
   author = {James Foadi and David Waterman and Rita Giordano and Kutumbarao Nidamarthi},
   year = {2025},
   note = {R package version 0.5.2},
-  url = {https://jfoadi.github.io/cry/},
+  url = {https://jfoadi.github.io/cry/}
 }
 
 @article{Wickham2014,
@@ -82,7 +83,7 @@
   volume={59},
   pages={1--23},
   year={2014},
-  doi = {10.18637/jss.v059.i10}
+  doi={10.18637/jss.v059.i10}
 }
 
 @article{Ku1966,
@@ -93,7 +94,8 @@
   number={4},
   pages={263},
   year={1966},
-  publisher={US Government Printing Office}
+  publisher={US Government Printing Office},
+  doi={10.6028/jres.070C.025}
 }
 
 @book{Emsley1998,
@@ -102,7 +104,6 @@
   publisher = "Oxford University Press",
   series    = "Oxford Chemistry Guides",
   edition   =  3,
-  month     =  mar,
   year      =  1998,
   address   = "London, England"
 }
@@ -116,21 +117,26 @@
   pages={186--197},
   year={2009},
   publisher={Wiley Online Library},
-  doi = {10.1002/chem.200800987}
+  doi={10.1002/chem.200800987}
 }
 
 @article{Jakob2025,
-  author = {Jakob, K. and Walsh, A. and Reuter, K. and Margraf, J. T.},
[…]
