+++ paper/paper.md
@@ -32,19 +32,19 @@ bibliography: paper.bib
 
 # Summary
 
-The Crystallographic Information File (CIF) is the standard format for disseminating crystal structure data, yet parsing and analyzing these files for large-scale computational and statistical analysis is a significant bottleneck for research in the chemical and material sciences [@Hall1991]. `crystract` is an R package designed to provide an efficient, open-source solution for the batch processing and statistical analysis of CIF files. The package streamlines the extraction of metadata, unit cell parameters, atomic coordinates, and symmetry operations for a singular or hundreds of CIF files at the same time. From the information stored in the CIF file, our package can determine the first-neighbor bonding environment around each symmetrically independent atom in a unit cell, with all interatomic distances and bond angles, while propagating experimental uncertainties. Furthermore, a comprehensive workflow for efficient extraction and processing of these data is provided, centered around the calculation of average interatomic distances, as identified as a key parameter for a streamlined comparison of structural features of different compounds. One of the most important features within this workflow is the package's ability to process positional or occupational disorder. To handle positional disorder, a filtering function to eliminate non-physical distances is provided. Occupational disorder is taken into account via the possibility of calculating an occupancy-weighted average interatomic distance. Additionally, filtering functions to calculate average distances, only including user-specified elements or atomic positions, are available. This paper outlines the architecture and core functionalities of `crystract`, demonstrating its utility with a practical example.
+The Crystallographic Information File (CIF) is the standard format for disseminating crystal structure data, yet parsing and analyzing these files for large-scale computational and statistical analysis is a significant bottleneck for research in the chemical and material sciences[@Hall1991]. `crystract` is an R package designed to provide an efficient, open-source solution for the batch processing and statistical analysis of CIF files. The package streamlines the extraction of metadata, unit cell parameters, atomic coordinates, and symmetry operations for a singular or hundreds of CIF files at the same time. From the information stored in the CIF file, our package can determine the first-neighbor bonding environment around each symmetrically independent atom in a unit cell, with all interatomic distances and bond angles, while propagating experimental uncertainties (i.e., the estimated standard deviations of the unit cell parameters and fractional atomic coordinates). Furthermore, a comprehensive workflow for efficient extraction and processing of these data is provided, centered around the calculation of average interatomic distances, as identified as a key parameter for a streamlined comparison of structural features of different compounds. One of the most important features within this workflow is the package's ability to process positional or occupational disorder. To handle positional disorder, a filtering function to eliminate non-physical distances is provided. Occupational disorder is taken into account via the possibility of calculating an occupancy-weighted average interatomic distance. Additionally, filtering functions to calculate average distances, only including user-specified elements or atomic positions, are available. This paper outlines the architecture and core functionalities of `crystract`, demonstrating its utility with a practical example.
 
 # Statement of need
 
[…]
