+++ packages/crystract/R/extract_cif_data.R
@@ -1,23 +1,24 @@
 #' @title Generic Value Extractor (Internal)
 #' @description Extracts a single value from CIF content based on a matching text pattern.
 #' @param cif_content A data.table containing the lines of a CIF file.
-#' @param pattern The text pattern (e.g., "_database_code_") to search for.
+#' @param patterns A character vector of text patterns to search for. The first matching pattern is returned.
 #' @param remove_pattern A boolean indicating whether to remove the search pattern from the result.
 #' @return A character string of the cleaned value, or NA if not found.
 #' @noRd
-extract_value <- function(cif_content, pattern, remove_pattern = TRUE) {
-  matching_lines <- cif_content[grepl(pattern, V1, fixed = TRUE)]
-  if (nrow(matching_lines) > 0) {
-    value <- matching_lines$V1[1]
-    if (remove_pattern) {
-      value <- gsub(pattern, "", value, fixed = TRUE)
+extract_value <- function(cif_content, patterns, remove_pattern = TRUE) {
+  for (pattern in patterns) {
+    matching_lines <- cif_content[grepl(pattern, V1, fixed = TRUE)]
+    if (nrow(matching_lines) > 0) {
+      value <- matching_lines$V1[1]
+      if (remove_pattern) {
+        value <- gsub(pattern, "", value, fixed = TRUE)
+      }
+      value <- gsub("'", "", value)
+      value <- trimws(value)
+      return(value)
     }
-    value <- gsub("'", "", value)
-    value <- trimws(value)
-    return(value)
-  } else {
-    return(NA)
   }
+  return(NA)
 }
 
 #' @title Extract Database Code from CIF Content
@@ -49,7 +50,15 @@ extract_database_code <- function (cif_content) {
 #'   extract_chemical_formula(cif_content)
 #' }
 extract_chemical_formula <- function(cif_content) {
-  extract_value(cif_content, "_chemical_formula_sum")
+  extract_value(
+    cif_content,
+    c(
+      "_chemical_formula_sum",
+      "_chemical_formula_structural",
+      "_chemical_formula_moiety",
+      "_chemical_formula_analytical"
+    )
+  )
 }
 
 #' @title Extract Structure Type from CIF Content
@@ -81,12 +90,14 @@ extract_structure_type <- function(cif_content) {
 #'   extract_space_group_name(cif_content)
 #' }
 extract_space_group_name <- function(cif_content) {
-  val <- extract_value(cif_content, "_space_group_name_H-M_alt")
-  if (is.na(val))
-    val <- extract_value(cif_content, "_symmetry_space_group_name_H-M")
-  if (is.na(val))
-    val <- extract_value(cif_content, "_space_group_name_H-M")
-  return(val)
+  extract_value(
+    cif_content,
+    c(
+      "_space_group_name_H-M_alt",
+      "_symmetry_space_group_name_H-M",
+      "_space_group_name_H-M"
+    )
+  )
 }
 
 #' @title Extract Space Group Number from CIF Content
@@ -102,10 +113,8 @@ extract_space_group_name <- function(cif_content) {
 #'   extract_space_group_number(cif_content)
 #' }
 extract_space_group_number <- function(cif_content) {
-  val <- extract_value(cif_content, "_space_group_IT_number")
-  if (is.na(val))
-    val <- extract_value(cif_content, "_symmetry_Int_Tables_number")
-  return(val)
+  extract_value(cif_content,
+                c("_space_group_IT_number", "_symmetry_Int_Tables_number"))
[…]

[…]
