### 2026-04-07 52239d7
Update Wyckoff examples from 4a to 4d
