+++ packages/crystract/doc/crystract.R
@@ -163,16 +163,16 @@ print("Atomic coordinates showing Wyckoff information:")
 print(atomic_coordinates[, .(Label, WyckoffSymbol, WyckoffMultiplicity)])
 cat("\n")
 
-# 2. Filter bonds where the central atom is on the "4a" Wyckoff site.
-bonds_from_4a_site <- filter_by_wyckoff_symbol(
+# 2. Filter bonds where the central atom is on the "4d" Wyckoff site.
+bonds_from_4d_site <- filter_by_wyckoff_symbol(
   data_table = bonded_pairs_with_error,
   atomic_coordinates = atomic_coordinates,
   atom_col = "Atom1",
-  wyckoff_symbols = "4a"  # Note: change to "a" if your package only expects the letter
+  wyckoff_symbols = "4d"
 )
 
 print(paste("Number of rows in original bond table:", nrow(bonded_pairs_with_error)))
-print(paste("Number of rows after filtering for site '4a':", nrow(bonds_from_4a_site)))
+print(paste("Number of rows after filtering for site '4d':", nrow(bonds_from_4d_site)))
 
 ## ----filter-ghost-distances, eval=TRUE-----------------------------------
 # A distance d is kept if: (r1+r2)*(1-margin) <= d <= (r1+r2)*(1+margin)
@@ -210,8 +210,8 @@ cat("Number of bonds after excluding 'Ge':", nrow(bonds_without_sn), "\n")
 # created in section 1.2.6.
 
 # Then, define the Wyckoff sites belonging to the network. 
-# For Okruginite, the framework atoms are on the "4a" sites.
-network_wyckoff_sites <- "4d" # Note: change to "a" if your package only expects the letter
+# For this file, the framework atoms are on the "4d" sites.
+network_wyckoff_sites <- "4d"
 
 # Apply the function to the table of identified bonds.
 weighted_avg_bond_dist <- calculate_weighted_average_network_distance(
@@ -220,7 +220,7 @@ weighted_avg_bond_dist <- calculate_weighted_average_network_distance(
     wyckoff_symbols = network_wyckoff_sites
 )
 
-cat("Weighted average network bond distance for the '4a' sites:", weighted_avg_bond_dist, "Å\n")
+cat("Weighted average network bond distance for the '4d' sites:", weighted_avg_bond_dist, "Å\n")
 
 ## ----custom-radii-and-batch, eval=FALSE----------------------------------
 # # --- 1. Define and set the custom radii dictionary ---
