+++ packages/crystract/doc/crystract.Rmd
@@ -357,16 +357,16 @@ print("Atomic coordinates showing Wyckoff information:")
 print(atomic_coordinates[, .(Label, WyckoffSymbol, WyckoffMultiplicity)])
 cat("\n")
 
-# 2. Filter bonds where the central atom is on the "4a" Wyckoff site.
-bonds_from_4a_site <- filter_by_wyckoff_symbol(
+# 2. Filter bonds where the central atom is on the "4d" Wyckoff site.
+bonds_from_4d_site <- filter_by_wyckoff_symbol(
   data_table = bonded_pairs_with_error,
   atomic_coordinates = atomic_coordinates,
   atom_col = "Atom1",
-  wyckoff_symbols = "4a"  # Note: change to "a" if your package only expects the letter
+  wyckoff_symbols = "4d"
 )
 
 print(paste("Number of rows in original bond table:", nrow(bonded_pairs_with_error)))
-print(paste("Number of rows after filtering for site '4a':", nrow(bonds_from_4a_site)))
+print(paste("Number of rows after filtering for site '4d':", nrow(bonds_from_4d_site)))
 ```
 
 ### 2.3 Filtering Ghost Distances Using Atomic Radii
@@ -456,8 +456,8 @@ This "Ghosts -\> Bonds -\> Average" sequence ensures the final metric is calcula
 # created in section 1.2.6.
 
 # Then, define the Wyckoff sites belonging to the network. 
-# For Okruginite, the framework atoms are on the "4a" sites.
-network_wyckoff_sites <- "4d" # Note: change to "a" if your package only expects the letter
+# For this file, the framework atoms are on the "4d" sites.
+network_wyckoff_sites <- "4d"
 
 # Apply the function to the table of identified bonds.
 weighted_avg_bond_dist <- calculate_weighted_average_network_distance(
@@ -466,7 +466,7 @@ weighted_avg_bond_dist <- calculate_weighted_average_network_distance(
     wyckoff_symbols = network_wyckoff_sites
 )
 
-cat("Weighted average network bond distance for the '4a' sites:", weighted_avg_bond_dist, "Å\n")
+cat("Weighted average network bond distance for the '4d' sites:", weighted_avg_bond_dist, "Å\n")
 ```
 
 ## 3. End-to-End Example: High-Throughput Batch Processing & CSV Extraction
