+++ articles/crystract.md
@@ -46,16 +46,16 @@ simultaneously.
 ``` r
 # IMPORTANT: Update this path to point to your own downloaded CIF file.
 # 1. Try to find the file in the installed package
-cif_path <- system.file("extdata", "1590946.cif", package = "crystract")
+cif_path <- system.file("extdata", "2405219.cif", package = "crystract")
 
 # 2. If that fails, look in the source directory
 if (cif_path == "") {
-  cif_path <- "../inst/extdata/1590946.cif"
+  cif_path <- "../inst/extdata/2405219.cif"
 }
 
 # 3. Final check to provide a clear error if both fail
 if (!file.exists(cif_path)) {
-  stop("Could not find 1590946.cif in installed package or inst/extdata/")
+  stop("Could not find 2405219.cif in installed package or inst/extdata/")
 }
 
 # Run the pipeline on our single example file, extracting multiple bonding types at once
@@ -68,9 +68,9 @@ analysis_results <- analyze_single_cif(
 # It's a single row containing all our results in nested data.tables.
 str(analysis_results, max.level = 2)
 #> Classes 'data.table' and 'data.frame':   1 obs. of  27 variables:
-#>  $ file_name              : chr "1590946.cif"
-#>  $ database_code          : chr "depnum_ccdc_archive CCDC 1590946"
-#>  $ chemical_formula       : chr "Si1 Sr2"
+#>  $ file_name              : chr "2405219.cif"
+#>  $ database_code          : chr "depnum_ccdc_archive CCDC 2405219"
+#>  $ chemical_formula       : chr "Ge Yb"
 #>  $ structure_type         : logi NA
 #>  $ space_group_name       : chr "P n m a"
 #>  $ space_group_number     : chr "62"
@@ -78,62 +78,62 @@ str(analysis_results, max.level = 2)
 #>   ..$ :Classes 'data.table' and 'data.frame':    1 obs. of  12 variables:
 #>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
 #>  $ atomic_coordinates     :List of 1
-#>   ..$ :Classes 'data.table' and 'data.frame':    3 obs. of  14 variables:
+#>   ..$ :Classes 'data.table' and 'data.frame':    2 obs. of  14 variables:
 #>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
 #>  $ symmetry_operations    :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  3 variables:
 #>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
 #>  $ transformed_coords     :List of 1
-#>   ..$ :Classes 'data.table' and 'data.frame':    12 obs. of  5 variables:
+#>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  5 variables:
 #>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
 #>  $ expanded_coords        :List of 1
-#>   ..$ :Classes 'data.table' and 'data.frame':    324 obs. of  8 variables:
+#>   ..$ :Classes 'data.table' and 'data.frame':    216 obs. of  8 variables:
 #>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
 #>  $ distances              :List of 1
-#>   ..$ :Classes 'data.table' and 'data.frame':    969 obs. of  6 variables:
[…]
