### 2026-04-07 abda7e8
Added example to show off errors
