+++ packages/crystract/doc/crystract.Rmd
@@ -144,7 +144,7 @@ print(head(symmetry_operations))
 
 Now we use the asymmetric atoms and symmetry operations to computationally build the full crystal structure. This is a two-step process.
 
-**Formula Context: Symmetry Operations**
+**Formula: Symmetry Operations**
 
 A symmetry operation is an affine transformation that maps an initial fractional coordinate $(x, y, z)$ to a new coordinate $(x', y', z')$. It consists of a rotation/reflection matrix $\mathbf{W}$ and a translation vector $\mathbf{w}$.
 
@@ -158,9 +158,9 @@ $$
 
 The `apply_symmetry_operations` function applies all of the crystal's symmetry operations to each atom in the asymmetric unit, generating the complete set of atoms within the primary unit cell. Note that it requires the unit cell metrics to resolve distance tolerances across boundaries.
 
-**Formula Context: Supercell Expansion**
+**Formula: Supercell Expansion**
 
-To find all nearest neighbors of an atom, we must also consider atoms in adjacent unit cells. The `expand_transformed_coords` function generates a **supercell** (in this case, 3x3x3) by translating the primary unit cell atoms by integer vectors $(i, j, k)$, where $i, j, k$ each range from -1 to 1. An atom at fractional coordinate $(x, y, z)$ generates new coordinates: 
+To find all nearest neighbors of an atom, we must also consider atoms in adjacent unit cells. The `expand_transformed_coords` function generates a **supercell** (in this case, 3x3x3) by translating the primary unit cell atoms by integer vectors $(i, j, k)$, where $i, j, k$ each range from -1 to 1. An atom at fractional coordinate $(x, y, z)$ generates new coordinates:
 
 $$
 (x_{exp}, y_{exp}, z_{exp}) = (x + i, y + j, z + k)
@@ -180,7 +180,7 @@ print(head(expanded_coords))
 
 #### 1.2.5 Calculating Interatomic Distances
 
-**Formula Context: Interatomic Distance in a Triclinic System**
+**Formula: Interatomic Distance in a Triclinic System**
 
 Because unit cell axes are not always orthogonal, the simple Pythagorean theorem is insufficient. The distance $d$ between two atoms at fractional coordinates $(x_{f1}, y_{f1}, z_{f1})$ and $(x_{f2}, y_{f2}, z_{f2})$ requires the full crystallographic distance formula:
 
@@ -205,15 +205,13 @@ print(head(distances[order(Distance)]))
 
 #### 1.2.6 Identifying Bonds and Neighbors
 
-**Formula Context: Bonding and Coordination Number**
+**Formula: Bonding and Coordination Number**
 
 Identifying which atoms are "bonded" is key to determining the **coordination number (CN)**. `crystract` implements several robust, geometric, and mathematical approaches:
 
--   **`minimum_distance`**: Defines a custom distance cutoff for each central atom $i$: 
-$$ 
-d_i^\text{cut} = (1+\delta) d_i^\text{min} 
-$$
-Here, $d_i^\text{min}$ is the shortest distance from atom $i$ to any other atom, and $\delta$ is a tolerance parameter (default is 0.1).
+-   **`minimum_distance`**: Defines a custom distance cutoff for each central atom $i$: $$ 
+    d_i^\text{cut} = (1+\delta) d_i^\text{min} 
+    $$ Here, $d_i^\text{min}$ is the shortest distance from atom $i$ to any other atom, and $\delta$ is a tolerance parameter (default is 0.1).
[…]
