+++ reference/export_analysis_to_csv.md
@@ -39,6 +39,6 @@ if (file.exists(cif_path)) {
 
   unlink(out_dir, recursive = TRUE)
 }
-#> Analysis successfully exported to: /tmp/RtmpAzF7E2/cif_csvs
+#> Analysis successfully exported to: /tmp/Rtmpvnd3nH/cif_csvs
 # }
 ```
