### 2026-04-07 #9529 @Don-Tat
Hey @bobleesj,

Thank you for the detailed feedback. I have addressed the remaining issues from your checklist:

- [x] DOI link was fixed to be `doi={10.1016/j.commatsci.2012.10.028}`

- [x] References for ICSD, COD, and AMCSD were added. References for the bonding algorithms were added as well.

- [x] CRAN installation has been added to [README](https://github.com/PrabhuLab/ml-crystals/blob/main/packages/crystract/README.md)?

- [x] Both the paper and vignette now reflect a more detailed description of what the error propagation is and what the experimental uncertainties are (and where they are found in our example CIF file from CCDC). In order to highlight, we switched our running example file from the strontium silicide (which did not contain any data on experimental uncertainties) to a different file from CCDC that did. This change, along with the more detailed explanations on error propagation, should improve the user’s understanding of this functionality. Thanks for suggesting this!

Please let me know if there is anything else you need to move forward with the review.
