### 2026-04-20 #9529 @Don-Tat
Hi @zhubonan, @singularitti, and @hisacro,

I wanted to check in on the review for crystract.

All previously discussed technical hurdles, including the installation process, have been resolved via our CRAN release. I have also finished addressing the remaining documentation and reference suggestions from @bobleesj, who has since confirmed that the updates look good.

Please let me know if there is any other information or further adjustments needed to help complete the remaining checklists.

Thank you for your time and expertise!

Best,

Don Ngo
