### 2026-05-02 #9529 @singularitti
Hey guys @zhubonan @hisacro @Don-Tat @anirudhprabhu, I have finished installing your package and run all tests you provided. They have passed. I also test it on some of my CIF files and they worked. Nice job! I think it's ready for publication. Can you double check your referecens? Are they fixed now? Do you consider migrating this pacakge to other languages as well?
