+++ articles/crystract.md
@@ -21,6 +21,7 @@ mathematical formulas they employ.
 First, we load the `crystract` package.
 
 ``` r
+
 library(crystract)
 library(data.table)
 ```
@@ -44,6 +45,7 @@ propagation. You can easily specify which bonding algorithms to run
 simultaneously.
 
 ``` r
+
 # IMPORTANT: Update this path to point to your own downloaded CIF file.
 # 1. Try to find the file in the installed package
 cif_path <- system.file("extdata", "2405219.cif", package = "crystract")
@@ -76,83 +78,83 @@ str(analysis_results, max.level = 2)
 #>  $ space_group_number     : chr "62"
 #>  $ unit_cell_metrics      :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    1 obs. of  12 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ atomic_coordinates     :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    2 obs. of  14 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ symmetry_operations    :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  3 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ transformed_coords     :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  5 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ expanded_coords        :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    216 obs. of  8 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ distances              :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    430 obs. of  6 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ bonds_minimum_distance :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    9 obs. of  8 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>   .. ..- attr(*, "sorted")= chr [1:3] "Atom1" "Atom2" "Distance"
 #>  $ cn_minimum_distance    :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    2 obs. of  3 variables:
 #>   .. ..- attr(*, "sorted")= chr "Atom"
-#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x564acbfe3110> 
 #>  $ angles_minimum_distance:List of 1
[…]
