+++ index.md
@@ -67,7 +67,7 @@ radii and choosing the most appropriate bonding algorithm.
 
 When invoking `analyze_cif_files(..., bonding_algorithms = c(...))`, we
 recommend choosing your target algorithm based on the chemical makeup of
-your structure and the electronegativity differences ($\Delta EN$) of
+your structure and the electronegativity differences ($`\Delta EN`$) of
 the expected bonds.
 
 ![Bonding Algorithm Selection
@@ -97,6 +97,7 @@ crystal structure included inside the package itself, demonstrating the
 exact data outputs you can expect.
 
 ``` r
+
 library(crystract)
 library(data.table)
 
@@ -131,8 +132,8 @@ parsed along with their experimental uncertainties (if available in the
 CIF).
 
 | \_cell_length_a | \_cell_length_b | \_cell_length_c | \_cell_angle_alpha | \_cell_angle_beta | \_cell_angle_gamma |
-|----------------:|----------------:|----------------:|-------------------:|------------------:|-------------------:|
-|            8.11 |            5.15 |            9.54 |                 90 |                90 |                 90 |
+|---:|---:|---:|---:|---:|---:|
+| 8.11 | 5.15 | 9.54 | 90 | 90 | 90 |
 
 Extracted Unit Cell Parameters (Å and Degrees)
 
@@ -179,45 +180,45 @@ Analysis Object and its nested tables.
 
 #### 1. Master Analysis Object
 
-| Column Name           | Data Type | Description                                                                      |
-|:----------------------|:----------|:---------------------------------------------------------------------------------|
-| `file_name`           | Character | The name of the processed CIF file.                                              |
-| `database_code`       | Character | The unique identifier from the source database.                                  |
-| `chemical_formula`    | Character | The chemical sum formula extracted from the CIF.                                 |
-| `structure_type`      | Character | The name of the structure type.                                                  |
-| `space_group_name`    | Character | Hermann-Mauguin space group symbol.                                              |
-| `space_group_number`  | Character | International Tables space group number.                                         |
-| `unit_cell_metrics`   | List (DT) | Nested table containing lattice parameters.                                      |
-| `atomic_coordinates`  | List (DT) | Nested table of primary asymmetric atoms.                                        |
-| `symmetry_operations` | List (DT) | Nested table of symmetry operators.                                              |
-| `transformed_coords`  | List (DT) | Nested table of the full unit cell atoms.                                        |
-| `expanded_coords`     | List (DT) | Nested table of the supercell (3x3x3) atoms.                                     |
-| `distances`           | List (DT) | Nested table of all calculated interatomic distances.                            |
-| `bonded_pairs_*`      | List (DT) | Nested table of bonds detected via requested methods (e.g. `_minimum_distance`). |
-| `neighbor_counts_*`   | List (DT) | Nested table of coordination numbers for requested methods.                      |
-| `bond_angles_*`       | List (DT) | Nested table of calculated bond angles for requested methods.                    |
+| Column Name | Data Type | Description |
+|:---|:---|:---|
+| `file_name` | Character | The name of the processed CIF file. |
+| `database_code` | Character | The unique identifier from the source database. |
+| `chemical_formula` | Character | The chemical sum formula extracted from the CIF. |
+| `structure_type` | Character | The name of the structure type. |
[…]
