### 2026-05-12 b2dbeac
Update bibliography and add author ORCID
