+++ paper/paper.bib
@@ -73,6 +73,7 @@
   author = {James Foadi and David Waterman and Rita Giordano and Kutumbarao Nidamarthi},
   year = {2025},
   note = {R package version 0.5.2},
+  doi={10.32614/cran.package.cry},
   url = {https://jfoadi.github.io/cry/}
 }
 
@@ -105,7 +106,8 @@
   series    = "Oxford Chemistry Guides",
   edition   =  3,
   year      =  1998,
-  address   = "London, England"
+  address   = "London, England",
+  isbn = "978-0198558187"
 }
 
 @article{Pyykko2009,
@@ -127,7 +129,7 @@
   volume = {38},
   number = {5},
   pages = {e14226},
-  doi = {https://doi.org/10.1002/adma.202514226},
+  doi = {10.1002/adma.202514226},
   url = {https://advanced.onlinelibrary.wiley.com/doi/abs/10.1002/adma.202514226},
   eprint = {https://advanced.onlinelibrary.wiley.com/doi/pdf/10.1002/adma.202514226},
   year = {2026}
@@ -190,13 +192,19 @@
   pages = {772--775}
 }
 
-@article{Bergerhoff1987,
-  title={Crystallographic databases},
-  author={Bergerhoff, G{\"u}nter and Brown, ID and Allen, F and others},
-  journal={International Union of Crystallography, Chester},
-  volume={360},
-  pages={77--95},
-  year={1987}
+@article{Zagorac5024,
+  author = "Zagorac, D. and M{\"{u}}ller, H. and Ruehl, S. and Zagorac, J. and Rehme, S.",
+  title = "{Recent developments in the Inorganic Crystal Structure Database: theoretical crystal structure data and related features}",
+  journal = "Journal of Applied Crystallography",
+  year = "2019",
+  volume = "52",
+  number = "5",
+  pages = "918--925",
+  month = "Oct",
+  doi = {10.1107/S160057671900997X},
+  url = {https://doi.org/10.1107/S160057671900997X},
+  abstract = {The Inorganic Crystal Structure Database (ICSD) is the world's largest database of fully evaluated and published crystal structure data, mostly obtained from experimental results. However, the purely experimental approach is no longer the only route to discover new compounds and structures. In the past few decades, numerous computational methods for simulating and predicting structures of inorganic solids have emerged, creating large numbers of theoretical crystal data. In order to take account of these new developments the scope of the ICSD was extended in 2017 to include theoretical structures which are published in peer-reviewed journals. Each theoretical structure has been carefully evaluated, and the resulting CIF has been extended and standardized. Furthermore, a first classification of theoretical data in the ICSD is presented, including additional categories used for comparison of experimental and theoretical information.},
+  keywords = {Inorganic Crystal Structure Database, ICSD, theoretical structures, standardization, classification},
 }
 
 @article{Downs2003,
@@ -237,14 +245,14 @@
   year={1966},
   publisher={De Gruyter},
   address={Berlin, Boston},
-  doi={https://doi.org/10.1515/9783111708003}
+  doi={10.1515/9783111708003}
 }
 
 @article{Hoppe1979,
 	title = {Effective coordination numbers ({ECoN}) and mean fictive ionic radii ({MEFIR})},
 	volume = {150},
 	url = {https://doi.org/10.1524/zkri.1979.150.14.23},
-	doi = {doi:10.1524/zkri.1979.150.14.23},
+	doi = {10.1524/zkri.1979.150.14.23},
 	number = {1-4},
 	urldate = {2026-02-24},
 	journal = {Zeitschrift für Kristallographie - Crystalline Materials},
[…]
