+++ authors.md
@@ -13,13 +13,13 @@
 ## Citation
 
 Ngo D, Prabhu A, Hubner J (2026). *crystract: Crystallographic
-Information File (CIF) Data Processing Tools*. R package version 1.0.0,
+Information File (CIF) Data Processing Tools*. R package version 1.0.1,
 <https://prabhulab.github.io/ml-crystals/>.
 
     @Manual{,
       title = {crystract: Crystallographic Information File (CIF) Data Processing Tools},
       author = {Don Ngo and Anirudh Prabhu and Julia-Maria Hubner},
       year = {2026},
-      note = {R package version 1.0.0},
+      note = {R package version 1.0.1},
       url = {https://prabhulab.github.io/ml-crystals/},
     }
