+++ articles/crystract.md
@@ -78,83 +78,83 @@ str(analysis_results, max.level = 2)
 #>  $ space_group_number     : chr "62"
 #>  $ unit_cell_metrics      :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    1 obs. of  12 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ atomic_coordinates     :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    2 obs. of  14 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ symmetry_operations    :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  3 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ transformed_coords     :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    8 obs. of  5 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ expanded_coords        :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    216 obs. of  8 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ distances              :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    430 obs. of  6 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ bonds_minimum_distance :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    9 obs. of  8 variables:
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>   .. ..- attr(*, "sorted")= chr [1:3] "Atom1" "Atom2" "Distance"
 #>  $ cn_minimum_distance    :List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    2 obs. of  3 variables:
 #>   .. ..- attr(*, "sorted")= chr "Atom"
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
+#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5589cbc74110> 
 #>  $ angles_minimum_distance:List of 1
 #>   ..$ :Classes 'data.table' and 'data.frame':    22 obs. of  5 variables:
 #>   .. ..- attr(*, "sorted")= chr [1:3] "CentralAtom" "Neighbor1" "Neighbor2"
-#>   .. ..- attr(*, ".internal.selfref")=<pointer: 0x5608de9cb110> 
[…]
